chore(endpoints): let fable rip on refactoring Endpoints
