+++ frontend/src/queries/schema.json
@@ -16161,6 +16161,49 @@
             },
             "type": "object"
         },
+        "EndpointBreakdownLimitExceededSignalExtra": {
+            "additionalProperties": false,
+            "properties": {
+                "breakdown_limit": {
+                    "type": "number"
+                },
+                "endpoint_name": {
+                    "type": "string"
+                }
+            },
+            "required": ["endpoint_name", "breakdown_limit"],
+            "type": "object"
+        },
+        "EndpointBreakdownLimitExceededSignalInput": {
+            "additionalProperties": false,
+            "properties": {
+                "description": {
+                    "type": "string"
+                },
+                "extra": {
+                    "$ref": "#/definitions/EndpointBreakdownLimitExceededSignalExtra"
+                },
+                "remediation": {
+                    "$ref": "#/definitions/SignalRemediation"
+                },
+                "source_id": {
+                    "type": "string"
+                },
+                "source_product": {
+                    "const": "endpoints",
+                    "type": "string"
+                },
+                "source_type": {
+                    "const": "endpoint_breakdown_limit_exceeded",
+                    "type": "string"
+                },
+                "weight": {
+                    "type": "number"
+                }
+            },
+            "required": ["description", "extra", "source_id", "source_product", "source_type", "weight"],
+            "type": "object"
+        },
         "EndpointExecutionFailedSignalExtra": {
             "additionalProperties": false,
             "properties": {
@@ -40209,6 +40252,9 @@
                     "$ref": "#/definitions/EndpointExecutionFailedSignalInput"
                 },
                 {
+                    "$ref": "#/definitions/EndpointBreakdownLimitExceededSignalInput"
+                },
+                {
                     "$ref": "#/definitions/PgAnalyzeIssueSignalInput"
                 },
                 {
@@ -40321,6 +40367,7 @@
                 "issue_reopened",
                 "issue_spiking",
                 "endpoint_execution_failed",
+                "endpoint_breakdown_limit_exceeded",
                 "cross_source_issue",
                 "alert_state_change",
                 "health_issue"
