+++ frontend/src/queries/schema/schema-signals.ts
@@ -28,6 +28,7 @@ export enum SignalSourceType {
     ISSUE_REOPENED = 'issue_reopened',
     ISSUE_SPIKING = 'issue_spiking',
     ENDPOINT_EXECUTION_FAILED = 'endpoint_execution_failed',
+    ENDPOINT_BREAKDOWN_LIMIT_EXCEEDED = 'endpoint_breakdown_limit_exceeded',
     CROSS_SOURCE_ISSUE = 'cross_source_issue',
     ALERT_STATE_CHANGE = 'alert_state_change',
     HEALTH_ISSUE = 'health_issue',
@@ -300,6 +301,19 @@ export interface EndpointExecutionFailedSignalInput extends SignalInputBase {
     extra: EndpointExecutionFailedSignalExtra
 }
 
+// Endpoint breakdown limit exceeded — the 'Other' bucket appeared in results
+
+export interface EndpointBreakdownLimitExceededSignalExtra extends SignalExtraBase {
+    endpoint_name: string
+    breakdown_limit: number
+}
+
+export interface EndpointBreakdownLimitExceededSignalInput extends SignalInputBase {
+    source_type: 'endpoint_breakdown_limit_exceeded'
+    source_product: 'endpoints'
+    extra: EndpointBreakdownLimitExceededSignalExtra
+}
+
 // Signals scout — cross-source findings emitted by the headless Signals scout harness.
 
 export interface SignalsScoutEvidenceEntry {
@@ -426,6 +440,7 @@ export type SignalInput =
     | ConversationsTicketSignalInput
     | ErrorTrackingSignalInput
     | EndpointExecutionFailedSignalInput
+    | EndpointBreakdownLimitExceededSignalInput
     | PgAnalyzeIssueSignalInput
     | SignalsScoutSignalInput
     | LogsAlertStateChangeSignalInput
