+++ posthog/temporal/data_modeling/run_workflow.py
@@ -72,8 +72,8 @@ from products.data_modeling.backend.models.modeling import DataWarehouseModelPat
 from products.data_warehouse.backend.data_load.create_table import create_table_from_saved_query
 from products.data_warehouse.backend.data_load.saved_query_service import a_pause_saved_query_schedule
 from products.data_warehouse.backend.s3 import ensure_bucket_exists, get_s3_client
-from products.endpoints.backend.rate_limit import set_endpoint_materialization_ready
-from products.endpoints.backend.services.endpoint_materialization_service import prepare_executable_query
+from products.endpoints.backend.rate_limit import update_materialization_ready_for_saved_query
+from products.endpoints.backend.services.materialization import prepare_executable_query
 from products.warehouse_sources.backend.models.table import DataWarehouseTable
 
 LOGGER = get_logger(__name__)
@@ -1596,7 +1596,7 @@ async def update_saved_query_status(
 
     if saved_query.origin == DataWarehouseSavedQuery.Origin.ENDPOINT:
         is_ready = status == DataWarehouseSavedQuery.Status.COMPLETED
-        await database_sync_to_async(set_endpoint_materialization_ready)(team_id, saved_query.name, is_ready)
+        await database_sync_to_async(update_materialization_ready_for_saved_query)(team_id, saved_query, is_ready)
 
 
 @dataclasses.dataclass
