+++ products/endpoints/backend/api.py
@@ -1,90 +1,47 @@
+"""HTTP layer for the Endpoints product.
+
+The viewset is intentionally thin: authentication/permissions, request parsing,
+and response serialization. Business logic lives in ``backend/services``:
+
+- ``services.crud``: create/update/destroy orchestration + activity logging
+- ``services.execution``: the /run path (materialized / inline / DuckLake)
+- ``services.materialization``: enable/disable/preview materialization
+- ``services.strategies``: HogQL vs insight query behavior
+- ``services.validation``: request payload validation
+"""
+
 import re
-import time
-import uuid
-import builtins
 import dataclasses
-from collections.abc import Iterator
-from datetime import timedelta
-from typing import Any, Optional, Union, cast
 
 from django.shortcuts import get_object_or_404
-from django.utils import timezone
 
-import structlog
-import posthoganalytics
-from asgiref.sync import async_to_sync
-from dateutil.parser import isoparse
 from django_filters.rest_framework import DjangoFilterBackend
 from drf_spectacular.utils import OpenApiParameter, extend_schema_view
-from loginas.utils import is_impersonated_session
-from pydantic import (
-    BaseModel,
-    ValidationError as PydanticValidationError,
-)
+from pydantic import ValidationError as PydanticValidationError
 from rest_framework import serializers, status, viewsets
-from rest_framework.exceptions import PermissionDenied, Throttled, ValidationError
+from rest_framework.exceptions import PermissionDenied, ValidationError
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.request import Request
 from rest_framework.response import Response
 
 from posthog.schema import (
-    DashboardFilter,
     EndpointLastExecutionTimesRequest,
-    EndpointRefreshMode,
     EndpointRequest,
     EndpointRunRequest,
-    HogQLQuery,
-    HogQLQueryModifiers,
-    HogQLVariable,
-    PropertyOperator,
-    QueryRequest,
     QueryStatus,
     QueryStatusResponse,
-    RefreshType,
 )
-
-from posthog.hogql import ast
-from posthog.hogql.context import HogQLContext
-from posthog.hogql.errors import ExposedHogQLError, QueryError, ResolutionError
-from posthog.hogql.parser import parse_expr, parse_select
-from posthog.hogql.printer import to_printed_hogql
-from posthog.hogql.printer.hogql import HogQLPrinter
-from posthog.hogql.printer.utils import print_prepared_ast
-from posthog.hogql.visitor import CloningVisitor
 
 from posthog.api.documentation import extend_schema
 from posthog.api.log_entries import LogEntryMixin
 from posthog.api.mixins import PydanticModelMixin, ValidatedRequest, validated_request
-from posthog.api.query import _process_query_request
 from posthog.api.routing import TeamAndOrgViewSetMixin
-from posthog.api.services.query import process_query_model
 from posthog.api.shared import UserBasicSerializer
-from posthog.api.tagged_item import TaggedItemViewSetMixin, cleanup_orphan_tags, set_tags_on_object
+from posthog.api.tagged_item import TaggedItemViewSetMixin
 from posthog.api.utils import action
 from posthog.auth import ProjectSecretAPIKeyAuthentication
-from posthog.clickhouse.client.connection import Workload
-from posthog.clickhouse.client.limit import ConcurrencyLimitExceeded
-from posthog.clickhouse.query_tagging import (
-    Feature,
-    Product,
-    get_query_tag_value,
-    is_api_key_access_method,
-    tag_queries,
-)
-from posthog.ducklake.common import get_duckgres_server_for_organization
-from posthog.errors import ExposedCHQueryError
-from posthog.event_usage import get_request_analytics_properties, report_user_action
+from posthog.clickhouse.query_tagging import Product
 from posthog.exceptions_capture import capture_exception
-from posthog.hogql_queries.insights.utils.breakdowns import BREAKDOWN_NULL_STRING_LABEL, BREAKDOWN_OTHER_STRING_LABEL
-from posthog.hogql_queries.query_runner import BLOCKING_EXECUTION_MODES
-from posthog.models import User
-from posthog.models.activity_logging.activity_log import (
-    ActivityContextBase,
-    Change,
-    Detail,
-    changes_between,
-    log_activity,
-)
 from posthog.permissions import (
     APIScopePermission,
     TeamMemberAccessPermission,
@@ -93,42 +50,9 @@ from posthog.permissions import (
 from posthog.rbac.access_control_api_mixin import AccessControlViewSetMixin
 from posthog.rbac.user_access_control import access_level_satisfied_for_resource
 from posthog.schema_migrations.upgrade import upgrade
-from posthog.synthetic_user import SyntheticUser
-from posthog.types import InsightQueryNode
 
-from products.data_modeling.backend.models.datawarehouse_saved_query import DataWarehouseSavedQuery
-from products.data_modeling.backend.services.saved_query_dag_sync import delete_node_from_dag, sync_saved_query_to_dag
-from products.data_warehouse.backend.data_load.saved_query_service import trigger_saved_query_schedule
-from products.endpoints.backend.insight_transformers import (
-    MaterializedSeriesMismatchError,
-    transform_materialized_insight_response,
-)
-from products.endpoints.backend.logs import ENDPOINTS_LOG_SOURCE, build_execution_message, log_endpoint_execution
-from products.endpoints.backend.materialization import (
-    ENDPOINT_BREAKDOWN_LIMIT,
-    SUPPORTED_BUCKET_FUNCTIONS,
-    VariablePlaceholderFinder,
-    _extract_aggregate_name,
-    analyze_variables_for_materialization,
-    build_endpoint_hogql,
-    convert_insight_query_to_hogql,
-    get_reaggregation,
-    prepare_insight_query_for_endpoint,
-    transform_query_for_materialization,
-    transform_select_for_materialized_table,
-)
-from products.endpoints.backend.metrics import (
-    ENDPOINT_CACHE_RESULT_TOTAL,
-    ENDPOINT_CONCURRENCY_REJECTED_TOTAL,
-    ENDPOINT_DUCKLAKE_FALLBACK_TOTAL,
-    ENDPOINT_EXECUTION_DURATION_SECONDS,
-    ENDPOINT_EXECUTION_TOTAL,
-    ENDPOINT_HOGQL_RESULT_ROWS,
-    ENDPOINT_MATERIALIZATION_EVENT_TOTAL,
-    ENDPOINT_MATERIALIZED_FRESHNESS_RATIO,
-    ENDPOINT_VALIDATION_ERROR_TOTAL,
-    query_kind_label,
-)
+from products.endpoints.backend.constants import ENDPOINT_NAME_REGEX
+from products.endpoints.backend.logs import ENDPOINTS_LOG_SOURCE
 from products.endpoints.backend.models import Endpoint, EndpointVersion
 from products.endpoints.backend.openapi import generate_openapi_spec
 from products.endpoints.backend.rate_limit import (
@@ -136,7 +60,6 @@ from products.endpoints.backend.rate_limit import (
     EndpointProjectSecretApiKeyTeamBurstThrottle,
     EndpointProjectSecretApiKeyTeamSustainedThrottle,
     EndpointSustainedThrottle,
-    clear_endpoint_materialization_cache,
 )
 from products.endpoints.backend.serializers import (
     EndpointMaterializationSerializer,
@@ -145,251 +68,17 @@ from products.endpoints.backend.serializers import (
     EndpointRunResponseSerializer,
     EndpointVersionResponseSerializer,
 )
-from products.product_analytics.backend.models.insight_variable import InsightVariable
-from products.warehouse_sources.backend.models.external_data_schema import sync_frequency_to_sync_frequency_interval
-
-from common.hogvm.python.utils import HogVMException
-
-DATA_FRESHNESS_BUCKETS: dict[int, str] = {
-    900: "15min",
-    1800: "30min",
-    3600: "1hour",
-    21600: "6hour",
-    43200: "12hour",
-    86400: "24hour",
-    604800: "7day",
-}
-VALID_DATA_FRESHNESS_SECONDS: frozenset[int] = frozenset(DATA_FRESHNESS_BUCKETS)
-DEFAULT_DATA_FRESHNESS_SECONDS = 86400
-
-ENDPOINT_NAME_REGEX = r"^[a-zA-Z][a-zA-Z0-9_-]{0,127}$"
-
-logger = structlog.get_logger(__name__)
-
-
-def _emit_endpoint_failure_signal(
-    team,
-    endpoint: "Endpoint",
-    exc: BaseException,
-    *,
-    materialized: bool,
-    version: int | None = None,
-    saved_query_id: Union[str, uuid.UUID, None] = None,
-    query_kind: str | None = None,
-    executed_sql: str | None = None,
-    saved_query_status: str | None = None,
-    saved_query_last_run_at: str | None = None,
-    saved_query_columns: dict | None = None,
-    endpoint_columns: list | None = None,
-) -> None:
-    """Fire a Signal when an endpoint execution fails, so the AI can reason about it later.
-
-    Fails silently — signal emission must never mask the underlying error.
-    """
-    from products.signals.backend.facade.api import emit_signal
-
-    try:
-        error_class = type(exc).__name__
-        error_msg = str(exc)
-        version_str = f" v{version}" if version else ""
-
-        if materialized:
-            execution_mode = "materialized"
-            context = (
-                f"The materialized table (saved_query_id={saved_query_id}) may be stale, missing, or have a schema mismatch. "
-                f"Check whether the materialization refresh completed successfully and whether the underlying query still produces valid columns."
-            )
-            if saved_query_status:
-                context += f"\nSaved query status: {saved_query_status}"
-            if saved_query_last_run_at:
-                context += f", last materialized at: {saved_query_last_run_at}"
-            if saved_query_columns:
-                context += f"\nMaterialized table columns: {saved_query_columns}"
-            if endpoint_columns:
-                context += f"\nEndpoint version columns: {endpoint_columns}"
-        else:
-            execution_mode = "inline"
-            context = (
-                f"The query is executed on-demand against live data. "
-                f"Common causes: invalid HogQL syntax, missing or renamed properties, query timeout, or incompatible variable overrides."
-            )
-
-        parts = [
-            f"Endpoint '{endpoint.name}'{version_str} failed during {execution_mode} execution.",
-            f"Error: {error_class}: {error_msg}",
-        ]
-        if query_kind:
-            parts.append(f"Query kind: {query_kind}")
-        if executed_sql:
-            parts.append(f"Executed HogQL: {executed_sql}")
-        parts.append(context)
-        parts.append(f"Endpoint path: {endpoint.endpoint_path}")
-        description = "\n".join(parts)
-
-        async_to_sync(emit_signal)(
-            team=team,
-            source_product="endpoints",
-            source_type="endpoint_execution_failed",
-            source_id=f"{team.id}:{endpoint.name}",
-            description=description,
-            weight=0.5,
-            extra={
-                "endpoint_name": endpoint.name,
-                "endpoint_version": version,
-                "materialized": materialized,
-                "saved_query_id": str(saved_query_id) if saved_query_id else None,
-                "error_class": error_class,
-                "error_message": error_msg,
-            },
-        )
-    except Exception:
-        logger.exception("Failed to emit endpoint failure signal", endpoint_name=endpoint.name)
-
-
-def _add_where_condition(select_query: ast.SelectQuery, condition: ast.Expr) -> None:
-    """Append a condition to select_query.where with AND."""
-    if select_query.where:
-        select_query.where = ast.And(exprs=[select_query.where, condition])
-    else:
-        select_query.where = condition
-
-
-def _get_single_breakdown_property(breakdown_filter: dict) -> str | None:
-    """Extract the breakdown property name from either legacy or new format.
-
-    Legacy: {"breakdown": "$browser", "breakdown_type": "event"}
-    New:    {"breakdowns": [{"property": "$browser", "type": "event"}]}
-    """
-    breakdown = breakdown_filter.get("breakdown")
-    if breakdown:
-        return breakdown
-
-    breakdowns = breakdown_filter.get("breakdowns") or []
-    if len(breakdowns) == 1:
-        return breakdowns[0].get("property")
-
-    return None
-
-
-def _iter_breakdowns(breakdown_filter: dict) -> Iterator[tuple[str, str]]:
-    """Yield (property_name, property_type) from legacy or new breakdown format."""
-    breakdown = breakdown_filter.get("breakdown")
-    if breakdown:
-        yield (breakdown, breakdown_filter.get("breakdown_type", "event"))
-        return
-    for b in breakdown_filter.get("breakdowns") or []:
-        prop = b.get("property")
-        if prop:
-            yield (prop, b.get("type", "event"))
-
-
-def _get_breakdown_properties(breakdown_filter: dict) -> list[str]:
-    """Extract all breakdown property names from either legacy or new format."""
-    return [name for name, _ in _iter_breakdowns(breakdown_filter)]
-
-
-def _get_breakdown_column_indices(breakdown_filter: dict) -> dict[str, int]:
-    """Map breakdown property names to their 1-based index in the breakdown_value array.
-
-    Single breakdown: {"$browser": 0}  (0 means use the whole array, not an index)
-    Multiple breakdowns: {"$browser": 1, "$os": 2}
-    """
-    props = _get_breakdown_properties(breakdown_filter)
-    if len(props) <= 1:
-        return {props[0]: 0} if props else {}
-    return {prop: i + 1 for i, prop in enumerate(props)}
-
-
-def _get_breakdown_infos(breakdown_filter: dict) -> list[tuple[str, str]]:
-    """Extract all breakdown property name and type pairs.
-
-    Returns list of (property_name, property_type) tuples.
-    """
-    return list(_iter_breakdowns(breakdown_filter))
-
-
-def _endpoint_refresh_mode_to_refresh_type(
-    mode: EndpointRefreshMode | None,
-) -> RefreshType:
-    """
-    Map EndpointRefreshMode to RefreshType.
-
-    - cache -> blocking
-    - force/direct -> force_blocking (materialization bypass handled in _should_use_materialized_table)
-    """
-    if mode is None or mode == EndpointRefreshMode.CACHE:
-        return RefreshType.BLOCKING
-    return RefreshType.FORCE_BLOCKING
-
-
-@dataclasses.dataclass(frozen=True)
-class EndpointContext(ActivityContextBase):
-    id: Optional[int] = None
-    version: Optional[int] = None
-
-
-@dataclasses.dataclass(frozen=True)
-class EndpointPagination:
-    limit: int
-    offset: int
-    ceiling: int | None
-
-    def apply_to(self, select_query: ast.SelectQuery) -> None:
-        """Set LIMIT and OFFSET on an AST SelectQuery for pagination."""
-        if self.ceiling is not None and self.offset >= self.ceiling:
-            select_query.limit = ast.Constant(value=0)
-            select_query.offset = ast.Constant(value=0)
-        else:
-            remaining = (self.ceiling - self.offset) if self.ceiling is not None else None
-            effective_limit = min(self.limit, remaining) if remaining is not None else self.limit
-            select_query.limit = ast.Constant(value=effective_limit + 1)
-            if self.offset > 0:
-                select_query.offset = ast.Constant(value=self.offset)
-
-    def process_results(self, result: dict) -> None:
-        """Trim the extra row and annotate the result dict with pagination metadata."""
-        rows = result.get("results", [])
-        has_more = len(rows) > self.limit
-        if has_more:
-            result["results"] = rows[: self.limit]
-        if self.ceiling is not None and self.offset + len(result["results"]) >= self.ceiling:
-            has_more = False
-        result["hasMore"] = has_more
-        result["limit"] = self.limit
-        result["offset"] = self.offset
-
-
-class _PlaceholderPreservingPrinter(HogQLPrinter):
-    """HogQL printer that preserves {placeholder} syntax instead of raising on unresolved placeholders."""
-
-    def visit_placeholder(self, node: ast.Placeholder) -> str:
-        if node.field is None:
-            raise QueryError("You can not use placeholders here")
-        return f"{{{node.field}}}"
-
-
-def _cant_materialize_response(reason: str) -> Response:
-    return Response(
-        {
-            "can_materialize": False,
-            "reason": reason,
-            "transformed_query": None,
-            "execution_query": None,
-            "range_pairs": [],
-            "aggregates": [],
-        }
-    )
-
-
-def _validate_bucket_overrides(bucket_overrides: dict[str, str] | None) -> None:
-    """Raise ValidationError if any bucket override value is not in SUPPORTED_BUCKET_FUNCTIONS."""
-    if not bucket_overrides:
-        return
-
-    invalid = {k: v for k, v in bucket_overrides.items() if v not in SUPPORTED_BUCKET_FUNCTIONS}
-    if invalid:
-        valid_options = list(SUPPORTED_BUCKET_FUNCTIONS.keys())
-        raise ValidationError(f"Invalid bucket override values: {invalid}. Valid options: {valid_options}")
+from products.endpoints.backend.services.crud import EndpointCrudService
+from products.endpoints.backend.services.execution import EndpointExecutionService
+from products.endpoints.backend.services.materialization import (
+    EndpointMaterializationService,
+    build_materialization_info,
+)
+from products.endpoints.backend.services.validation import (
+    validate_bucket_overrides,
+    validate_endpoint_request,
+    validate_update_request,
+)
 
 
 class MaterializationPreviewRequestSerializer(serializers.Serializer):
@@ -400,18 +89,6 @@ class MaterializationPreviewRequestSerializer(serializers.Serializer):
         allow_null=True,
         help_text='Per-column bucket function overrides, e.g. {"timestamp": "hour"}',
     )
-
-
-class _RunRejected(Exception):
-    """Internal control-flow signal: a run request was rejected during validation.
-
-    Carries the 4xx Response to return to the caller. Caught in `run()`, where the rejection is
-    logged once — alongside DRF ValidationErrors — so every rejected run surfaces in the logs.
-    """
-
-    def __init__(self, response: Response):
-        self.response = response
-        super().__init__()
 
 
 @extend_schema_view(
@@ -431,7 +108,6 @@ class EndpointViewSet(
 ):
     authentication_classes = [ProjectSecretAPIKeyAuthentication]
     psak_allowed_actions = ["run"]
-    # NOTE: Do we need to override the scopes for the "create"
     scope_object = "endpoint"
     # Read endpoint execution logs from the `log_entries` table keyed by this source.
     log_source = ENDPOINTS_LOG_SOURCE
@@ -482,6 +158,10 @@ class EndpointViewSet(
             return [permission() for permission in (IsAuthenticated, APIScopePermission, TeamMemberAccessPermission)]
         raise NotImplementedError()
 
+    # ------------------------------------------------------------------
+    # Access control helpers
+    # ------------------------------------------------------------------
+
     def _get_endpoint_with_object_access(self, name: str | None) -> Endpoint:
         """Fetch a team-scoped endpoint by name and enforce object-level access controls.
 
@@ -500,14 +180,16 @@ class EndpointViewSet(
         defaults so a restrictive project default doesn't break the public execution API. Creators
         and org admins always pass (specific_access_level_for_object returns the highest level).
         """
-        # PSAK scopes are project-wide within their resource type and deliberately ignore
-        # object-level access controls — the key's team binding and `endpoint:read` scope are
-        # the whole authorization story (see posthog/scopes.py).
+        # PSAK scopes are project-wide and deliberately bypass object-level access controls.
         if is_authenticated_via_project_secret_api_key(self.request):
             return
         specific_level = self.user_access_control.specific_access_level_for_object(endpoint)
         if specific_level is not None and not access_level_satisfied_for_resource("endpoint", specific_level, "viewer"):
             raise PermissionDenied("You do not have access to this endpoint.")
+
+    # ------------------------------------------------------------------
+    # Request parsing helpers
+    # ------------------------------------------------------------------
 
     def _parse_version_param(self, request: Request) -> int | None:
         """Parse version number from request body or query params.
@@ -531,56 +213,48 @@ class EndpointViewSet(
 
         return None
 
+    @staticmethod
+    def _parse_int_param(
+        body_value: int | None, query_param: str | None, name: str, min_value: int | None = None
+    ) -> tuple[int | None, Response | None]:
+        """Parse an integer from the request body or query params. Returns (value, error_response)."""
+        value = body_value
+        if value is None and query_param is not None:
+            try:
+                value = int(query_param)
+                if min_value is not None and value < min_value:
+                    raise ValueError()
+            except (ValueError, TypeError):
+                return None, Response(
+                    {"error": f"Invalid {name} parameter: {query_param}"},
+                    status=status.HTTP_400_BAD_REQUEST,
+                )
+        elif value is not None and min_value is not None and value < min_value:
+            return None, Response(
+                {"error": f"Invalid {name} parameter: {value}"},
+                status=status.HTTP_400_BAD_REQUEST,
+            )
+        return value, None
+
+    def _resolve_version_param(self, request: Request, endpoint: Endpoint) -> EndpointVersion | Response:
+        """Resolve the ?version=N param to a version, or a 404 Response if it doesn't exist."""
+        version_number = self._parse_version_param(request)
+        if version_number is None:
+            return endpoint.get_version()
+        try:
+            return endpoint.get_version(version_number)
+        except EndpointVersion.DoesNotExist:
+            return Response({"error": f"Version {version_number} not found"}, status=status.HTTP_404_NOT_FOUND)
+
+    # ------------------------------------------------------------------
+    # Serialization
+    # ------------------------------------------------------------------
+
     def _get_tag_names(self, endpoint: Endpoint) -> list[str]:
         """Read tag names from the prefetched cache or from the DB if not prefetched."""
         if hasattr(endpoint, "prefetched_tags"):
             return sorted({ti.tag.name for ti in endpoint.prefetched_tags})
         return sorted(endpoint.tagged_items.values_list("tag__name", flat=True))
-
-    def _apply_tags(self, endpoint: Endpoint, tags: list[str] | None) -> None:
-        """Replace the endpoint's tags. No-op when tags is None (field omitted)."""
-        if tags is None:
-            return
-        # `prefetched_tags` is a dynamic attribute populated by the TaggedItem prefetch / set_tags_on_object;
-        # it's not declared on the Endpoint model.
-        endpoint.prefetched_tags = set_tags_on_object(tags, endpoint)  # type: ignore[attr-defined]
-        cleanup_orphan_tags(endpoint.team_id)
-
-    @extend_schema(exclude=True)
-    @action(methods=["POST"], detail=False)
-    def bulk_update_tags(self, request: Request, *args, **kwargs) -> Response:
-        # The inherited TaggedItemViewSetMixin.bulk_update_tags assumes integer PKs (its
-        # BulkUpdateTagsRequestSerializer validates ids as IntegerField). Endpoint uses
-        # UUID PKs, so the action is unusable. Return 405 until the mixin gains UUID support.
-        return Response(
-            {"detail": "Bulk tag updates are not supported for endpoints."},
-            status=status.HTTP_405_METHOD_NOT_ALLOWED,
-        )
-
-    def _build_materialization_info(self, version: EndpointVersion, endpoint_name: str | None = None) -> dict:
-        """Build materialization status dict for a version."""
-        if version.saved_query:
-            result = {
-                "status": version.saved_query.status or "Unknown",
-                "can_materialize": True,
-                "last_materialized_at": (
-                    version.saved_query.last_run_at.isoformat() if version.saved_query.last_run_at else None
-                ),
-                "error": (version.saved_query.latest_error or "")
-                if version.saved_query.status != DataWarehouseSavedQuery.Status.COMPLETED
-                else "",
-                "saved_query_id": str(version.saved_query.id),
-            }
-        else:
-            can_mat, reason = version.can_materialize()
-            result = {
-                "can_materialize": can_mat,
-                "reason": reason if not can_mat else None,
-            }
-
-        if endpoint_name is not None:
-            result["name"] = endpoint_name
-        return result
 
     def _serialize(
         self,
@@ -624,7 +298,7 @@ class EndpointViewSet(
             "versions_count": endpoint.versions.count(),
             "derived_from_insight": endpoint.derived_from_insight,
             "last_executed_at": endpoint.last_executed_at.isoformat() if endpoint.last_executed_at else None,
-            "materialization": self._build_materialization_info(version),
+            "materialization": build_materialization_info(version),
             "bucket_overrides": version.bucket_overrides,
             "columns": version.get_columns() if version else [],
             "tags": self._get_tag_names(endpoint),
@@ -641,6 +315,10 @@ class EndpointViewSet(
             result["version_created_by"] = UserBasicSerializer(version.created_by).data if version.created_by else None
 
         return result
+
+    # ------------------------------------------------------------------
+    # CRUD actions
+    # ------------------------------------------------------------------
 
     @extend_schema(
         responses={200: EndpointResponseSerializer(many=True)},
@@ -681,176 +359,6 @@ class EndpointViewSet(
             capture_exception(e, {"endpoint_name": endpoint.name, "team_id": self.team_id})
             raise ValidationError("Failed to retrieve endpoint.")
 
-    def _validate_data_freshness(self, data_freshness_seconds: int | None) -> None:
-        """Validate data_freshness_seconds is one of the allowed bucket values."""
-        if data_freshness_seconds is None:
-            return
-        if data_freshness_seconds not in VALID_DATA_FRESHNESS_SECONDS:
-            allowed = sorted(VALID_DATA_FRESHNESS_SECONDS)
-            raise ValidationError(
-                {
-                    "data_freshness_seconds": (
-                        f"Data freshness must be one of: {allowed} seconds "
-                        "(15 minutes, 30 minutes, 1 hour, 6 hours, 12 hours, 24 hours, 7 days)."
-                    )
-                }
-            )
-
-    def _validate_hogql_query(self, query: HogQLQuery) -> None:
-        """Validate that a HogQL query can be parsed and the variables are valid."""
-        try:
-            ast_node = parse_select(query.query)
-        except ExposedHogQLError as e:
-            raise ValidationError({"query": f"Invalid HogQL query: {e}"}) from e
-        except ResolutionError as e:
-            capture_exception(e)
-            raise ValidationError({"query": "Invalid HogQL query: unable to resolve table or field references."})
-        except Exception as e:
-            capture_exception(e)
-            raise ValidationError({"query": "Unknown error occurred parsing the query."})
-
-        self._validate_variable_placeholders(ast_node, query.variables or {})
-
-    def _sync_hogql_query_variables(self, query: HogQLQuery) -> None:
-        """Sync query variable definitions with the placeholders used in the query string."""
-        try:
-            ast_node = parse_select(query.query)
-        except Exception:
-            # Validation runs separately and will surface parse errors.
-            return
-
-        finder = VariablePlaceholderFinder()
-        finder.visit(ast_node)
-
-        placeholder_names = {str(p.chain[1]) for p in finder.variable_placeholders if p.chain and len(p.chain) > 1}
-        if not placeholder_names:
-            query.variables = None
-            return
-
-        existing_variables = query.variables or {}
-        existing_variable_ids: list[str] = []
-        for variable in existing_variables.values():
-            if not variable.variableId:
-                continue
-            try:
-                uuid.UUID(variable.variableId)
-            except ValueError:
-                continue
-            existing_variable_ids.append(variable.variableId)
-
-        team_variables = InsightVariable.objects.filter(team=self.team, id__in=existing_variable_ids)
-        team_variables_by_id = {str(variable.id): variable for variable in team_variables}
-
-        synced_variables: dict[str, HogQLVariable] = {}
-        existing_code_names: set[str] = set()
-
-        for variable_id, variable in existing_variables.items():
-            if variable.code_name and variable.code_name in placeholder_names:
-                if variable.value is None and variable.isNull is not True:
-                    team_variable = team_variables_by_id.get(variable.variableId)
-                    if team_variable:
-                        variable.value = team_variable.default_value
-                        variable.isNull = team_variable.default_value is None
-                synced_variables[variable_id] = variable
-                existing_code_names.add(variable.code_name)
-
-        missing_code_names = placeholder_names - existing_code_names
-        if missing_code_names:
-            missing_variables = InsightVariable.objects.filter(team=self.team, code_name__in=missing_code_names)
-
-            for variable in missing_variables:
-                if not variable.code_name:
-                    continue
-                synced_variables[str(variable.id)] = HogQLVariable(
-                    variableId=str(variable.id),
-                    code_name=variable.code_name,
-                    value=variable.default_value,
-                    isNull=variable.default_value is None,
-                )
-
-        query.variables = synced_variables or None
-
-    def _validate_variable_placeholders(self, node: ast.AST, variables: Optional[dict[str, HogQLVariable]]) -> None:
-        """Validate that every {variables.X} placeholder in the query has a matching variable definition."""
-        finder = VariablePlaceholderFinder()
-        finder.visit(node)
-
-        if not finder.variable_placeholders:
-            return
-
-        placeholder_names = {str(p.chain[1]) for p in finder.variable_placeholders if p.chain and len(p.chain) > 1}
-
-        defined_code_names: set[str] = set()
-        if variables:
-            defined_code_names = {v.code_name for v in variables.values() if v.code_name}
-            variable_ids = set(variables.keys())
-
-        undefined: list[str] = sorted(placeholder_names - defined_code_names)
-        if undefined:
-            raise ValidationError(
-                {
-                    "query": f"Query references undefined variable(s): {', '.join(undefined)}. "
-                    "See https://posthog.com/docs/endpoints/variables for detail."
-                }
-            )
-
-        if variable_ids:
-            valid_uuids: set[str] = set()
-            invalid_uuids: set[str] = set()
-            invalid_ids: set[str] = set()
-            for vid in variable_ids:
-                try:
-                    uuid.UUID(vid)
-                    valid_uuids.add(vid)
-                except ValueError:
-                    invalid_uuids.add(vid)
-
-            if invalid_uuids:
-                raise ValidationError(
-                    {"query": f"Variable ID(s) not valid UUIDs: {', '.join(sorted(invalid_uuids))}. "}
-                )
-
-            if valid_uuids:
-                existing_ids = {
-                    str(id)
-                    for id in InsightVariable.objects.filter(team=self.team, id__in=valid_uuids).values_list(
-                        "id", flat=True
-                    )
-                }
-                invalid_ids = valid_uuids - existing_ids
-
-            if invalid_ids:
-                raise ValidationError(
-                    {
-                        "query": f"Variable ID(s) not found: {', '.join(sorted(invalid_ids))}. "
-                        "Make sure the variables exist in https://app.posthog.com/data-management/variables."
-                    }
-                )
-
-    def validate_request(self, data: EndpointRequest, strict: bool = True) -> None:
-        query = data.query
-        if not query and strict:
-            raise ValidationError({"query": "This field is required."})
-
-        name = data.name
-        if not name:
-            if name is not None or strict:
-                raise ValidationError({"name": "This field is required."})
-            return
-        if not isinstance(name, str) or not re.fullmatch(ENDPOINT_NAME_REGEX, name):
-            raise ValidationError(
-                {
-                    "name": f"Invalid name '{name}'. Must start with a letter, contain only alphanumeric characters, "
-                    "hyphens, or underscores, and be between 1 and 128 characters long."
-                }
-            )
-
-        if query and isinstance(query, HogQLQuery) and query.query:
-            self._sync_hogql_query_variables(query)
-            self._validate_hogql_query(query)
-
-        self._validate_data_freshness(data.data_freshness_seconds)
-
     @extend_schema(
         request=EndpointRequestSerializer,
         responses={201: EndpointResponseSerializer},
@@ -860,99 +368,14 @@ class EndpointViewSet(
         """Create a new endpoint."""
         upgraded_query = upgrade(request.data)
         data = self.get_model(upgraded_query, EndpointRequest)
-        self.validate_request(data, strict=True)
+        validate_endpoint_request(data, self.team, strict=True)
 
-        if Endpoint.objects.filter(team=self.team, name=data.name, deleted=False).exists():
-            raise ValidationError({"name": "An endpoint with this name already exists for this team."})
+        endpoint = EndpointCrudService(self.team, request).create(data)
 
-        try:
-            query_dict = cast(Union[HogQLQuery, InsightQueryNode], data.query).model_dump()
-            endpoint = Endpoint.objects.create(
-                team=self.team,
-                created_by=cast(User, request.user),
-                name=cast(str, data.name),  # verified in validate_request
-                is_active=data.is_active if data.is_active is not None else True,
-                current_version=1,
-                derived_from_insight=data.derived_from_insight,
-            )
-
-            try:
-                columns: list[dict] | None = EndpointVersion.extract_columns(query_dict, team_id=self.team.pk)
-            except Exception:
-                columns = None
-            EndpointVersion.objects.create(
-                endpoint=endpoint,
-                team=self.team,
-                version=1,
-                query=query_dict,
-                description=data.description or "",
-                data_freshness_seconds=(
-                    data.data_freshness_seconds
-                    if data.data_freshness_seconds is not None
-                    else DEFAULT_DATA_FRESHNESS_SECONDS
-                ),
-                created_by=cast(User, request.user),
-                columns=columns,
-            )
-
-            self._apply_tags(endpoint, data.tags)
-
-            log_activity(
-                organization_id=self.organization.id,
-                team_id=self.team.id,
-                user=cast(User, request.user),
-                was_impersonated=is_impersonated_session(request),
-                item_id=str(endpoint.id),
-                scope="Endpoint",
-                activity="created",
-                detail=Detail(name=endpoint.name),
-            )
-
-            report_user_action(
-                user=cast(User, request.user),
-                event="endpoint created",
-                properties={
-                    "endpoint_id": str(endpoint.id),
-                    "endpoint_name": endpoint.name,
-                    "query_kind": query_dict.get("kind") if isinstance(query_dict, dict) else None,
-                },
-                team=self.team,
-                request=request,
-            )
-
-            return Response(
-                self._serialize(endpoint, request),
-                status=status.HTTP_201_CREATED,
-            )
-
-        except Exception as e:
-            capture_exception(
-                e,
-                {
-                    "product": Product.ENDPOINTS,
-                    "team_id": self.team_id,
-                    "endpoint_name": data.name,
-                },
-            )
-            raise ValidationError("Failed to create endpoint.")
-
-    def validate_update_request(
-        self,
-        data: EndpointRequest,
-        endpoint: Endpoint | None = None,
-        strict: bool = True,
-    ) -> None:
-        self._validate_data_freshness(data.data_freshness_seconds)
-
-        # Determine final states after this request (for validation)
-        will_be_active = data.is_active if data.is_active is not None else (endpoint.is_active if endpoint else True)
-
-        if not will_be_active and data.is_materialized is True:
-            raise ValidationError({"is_materialized": "Cannot enable materialization on inactive endpoint."})
-
-        if data.query and isinstance(data.query, HogQLQuery) and data.query.query:
-            self._sync_hogql_query_variables(data.query)
-            self._validate_hogql_query(data.query)
+        return Response(
+            self._serialize(endpoint, request),
+            status=status.HTTP_201_CREATED,
+        )
 
     @extend_schema(
         request=EndpointRequestSerializer,
@@ -969,7 +392,6 @@ class EndpointViewSet(
         # Enforce object-level RBAC: a user with global editor scope can still be restricted
         # from a specific endpoint via per-object access controls.
         endpoint = self._get_endpoint_with_object_access(name)
-        endpoint_before_update = Endpoint.objects.get(pk=endpoint.id)
 
         upgraded_query = upgrade(request.data)
         data = self.get_model(upgraded_query, EndpointRequest)
@@ -979,1371 +401,49 @@ class EndpointViewSet(
             self.destroy(request, name=name)
             return Response({"success": True}, status=status.HTTP_200_OK)
 
-        self.validate_update_request(data, endpoint=endpoint, strict=False)
+        validate_update_request(data, self.team, endpoint=endpoint)
 
         version_number = self._parse_version_param(request)
-        target_version_override = None
-        if version_number is not None:
-            try:
-                target_version_override = endpoint.get_version(version_number)
-            except EndpointVersion.DoesNotExist:
-                raise ValidationError({"version": f"Version {version_number} not found for this endpoint."})
-            if data.query is not None:
-                raise ValidationError(
-                    {
-                        "query": "Cannot change query when targeting a specific version. Query changes create a new version."
-                    }
-                )
+        outcome = EndpointCrudService(self.team, request).update(endpoint, data, request.data, version_number)
 
-        try:
-            current_version = endpoint.get_version()
-            target_version = target_version_override or current_version
+        # When targeting a specific version, return version data; otherwise return endpoint data
+        if outcome.version_targeted and outcome.target_version is not None:
+            result = self._serialize(outcome.target_version)
+        else:
+            result = self._serialize(outcome.endpoint, request)
 
-            version_before_update = EndpointVersion.objects.get(pk=target_version.pk) if target_version else None
-            version_was_created = False
-            query_changed = False
-            new_query_dict = None
-            if data.query is not None:
-                new_query_dict = data.query.model_dump()
-                query_changed = endpoint.has_query_changed(new_query_dict)
+        if outcome.materialization_error:
+            result["materialization_error"] = outcome.materialization_error
 
-            # Deactivates the whole endpoint - we deactivate a version later if requested
-            if data.is_active is not None and target_version_override is None:
-                endpoint.is_active = data.is_active
-            endpoint.save()
-
-            final_is_active = data.is_active if data.is_active is not None else endpoint.is_active
-            was_materialized = current_version.saved_query_id is not None
-
-            # Step 1: Handle deactivation (disables materialization, prevents any materialization operations)
-            if not final_is_active and was_materialized:
-                self._disable_materialization(endpoint, request, current_version)
-
-            # Step 2: Handle query changes and versioning (independent of active/materialization state)
-            old_bucket_overrides: dict[str, str] | None = None
-            if query_changed and new_query_dict is not None:
-                if was_materialized:
-                    old_bucket_overrides = current_version.bucket_overrides
-
-                new_version = endpoint.create_new_version(query=new_query_dict, user=cast(User, request.user))
-                version_was_created = True
-                current_version = new_version
-                target_version = new_version
-
-            # Step 3: Update version-level fields on target version
-            if target_version:
-                update_fields = []
-                if data.description is not None:
-                    target_version.description = data.description
-                    update_fields.append("description")
-                if "data_freshness_seconds" in request.data:
-                    target_version.data_freshness_seconds = (
-                        data.data_freshness_seconds
-                        if data.data_freshness_seconds is not None
-                        else DEFAULT_DATA_FRESHNESS_SECONDS
-                    )
-                    update_fields.append("data_freshness_seconds")
-                # When targeting a specific version, is_active updates the version
-                if data.is_active is not None and target_version_override is not None:
-                    target_version.is_active = data.is_active
-                    update_fields.append("is_active")
-                if update_fields:
-                    update_fields.append("updated_at")
-                    target_version.save(update_fields=update_fields)
-
-            # Step 4: Handle materialization state (only if endpoint should be active)
-            stored_bucket_overrides = target_version.bucket_overrides if target_version else None
-            materialization_error: str | None = None
-            if final_is_active and target_version:
-                # When targeting a specific version, check that version's materialization state
-                # Otherwise use was_materialized (state before this update) to support materialization transfer
-                if target_version_override is not None:
-                    check_was_materialized = target_version.saved_query_id is not None
-                else:
-                    check_was_materialized = was_materialized
-
-                should_enable = data.is_materialized is True or (
-                    data.is_materialized is None and check_was_materialized
-                )
-                should_disable = data.is_materialized is False
-
-                if should_enable:
-                    bucket_overrides = request.data.get("bucket_overrides")
-                    if bucket_overrides is None and version_was_created:
-                        bucket_overrides = old_bucket_overrides
-                    _validate_bucket_overrides(bucket_overrides)
-                    try:
-                        self._enable_materialization(
-                            endpoint,
-                            target_version.data_freshness_seconds,
-                            request,
-                            target_version,
-                            bucket_overrides=bucket_overrides,
-                        )
-                        # Trigger immediate refresh when bucket_overrides changed
-                        if (
-                            bucket_overrides is not None
-                            and bucket_overrides != stored_bucket_overrides
-                            and target_version.saved_query is not None
-                        ):
-                            try:
-                                trigger_saved_query_schedule(target_version.saved_query)
-                            except Exception:
-                                logger.warning(
-                                    "failed_to_trigger_materialization_refresh",
-                                    team_id=self.team_id,
-                                    saved_query_id=str(target_version.saved_query_id),
-                                )
-                    except Exception as e:
-                        if version_was_created:
-                            # The new version was already committed — don't fail the whole update.
-                            # Materialization can be retried via a subsequent update.
-                            materialization_error = str(e)
-                            logger.exception(
-                                "Materialization failed after version creation",
-                                endpoint_name=endpoint.name,
-                                version=target_version.version,
-                            )
-                            capture_exception(
-                                e,
-                                {
-                                    "product": Product.ENDPOINTS,
-                                    "team_id": self.team_id,
-                                    "endpoint_name": endpoint.name,
-                                    "version": target_version.version,
-                                },
-                            )
-                        else:
-                            raise
-                elif should_disable:
-                    self._disable_materialization(endpoint, request, target_version)
-
-            self._apply_tags(endpoint, data.tags)
-
-            endpoint_changes = changes_between("Endpoint", previous=endpoint_before_update, current=endpoint)
-            if endpoint_changes:
-                # endpoint-level activity
-                log_activity(
-                    organization_id=self.organization.id,
-                    team_id=self.team.id,
-                    user=cast(User, request.user),
-                    was_impersonated=is_impersonated_session(request),
-                    item_id=str(endpoint.id),
-                    scope="Endpoint",
-                    activity="updated",
-                    detail=Detail(name=endpoint.name, changes=endpoint_changes),
-                )
-
-            # TODO: Should these go to finally: ?
-            # version-level activity
-            if version_was_created:
-                query_change = Change(
-                    type="EndpointVersion",
-                    action="changed",
-                    field="query",
-                    before=version_before_update.query if version_before_update else None,
-                    after=target_version.query,
-                )
-                log_activity(
-                    organization_id=self.organization.id,
-                    team_id=self.team.id,
-                    user=cast(User, request.user),
-                    was_impersonated=is_impersonated_session(request),
-                    item_id=str(endpoint.id),
-                    scope="Endpoint",
-                    activity="version_created",
-                    detail=Detail(
-                        name=endpoint.name,
-                        changes=[query_change],
-                        context=EndpointContext(version=target_version.version),
-                    ),
-                )
-            elif target_version and version_before_update:
-                version_changes = changes_between(
-                    "EndpointVersion", previous=version_before_update, current=target_version
-                )
-
-                if version_changes:
-                    log_activity(
-                        organization_id=self.organization.id,
-                        team_id=self.team.id,
-                        user=cast(User, request.user),
-                        was_impersonated=is_impersonated_session(request),
-                        item_id=str(endpoint.id),
-                        scope="EndpointVersion",
-                        activity="version_updated",
-                        detail=Detail(
-                            name=endpoint.name,
-                            changes=version_changes,
-                            context=EndpointContext(version=target_version.version),
-                        ),
-                    )
-
-            # When targeting a specific version, return version data; otherwise return endpoint data
-            if target_version_override is not None:
-                result = self._serialize(target_version)
-            else:
-                result = self._serialize(endpoint, request)
-
-            if materialization_error:
-                result["materialization_error"] = materialization_error
-
-            return Response(result)
-
-        except ValidationError:
-            raise
-        except Exception as e:
-            current_version = endpoint.get_version()
-            capture_exception(
-                e,
-                {
-                    "product": Product.ENDPOINTS,
-                    "team_id": self.team_id,
-                    "endpoint_id": endpoint.id,
-                    "saved_query_id": current_version.saved_query.id if current_version.saved_query else None,
-                },
-            )
-            raise ValidationError("Failed to update endpoint.")
-
-    def _enable_materialization(
-        self,
-        endpoint: Endpoint,
-        data_freshness_seconds: int,
-        request: Request,
-        version: EndpointVersion | None = None,
-        bucket_overrides: dict[str, str] | None = None,
-    ) -> None:
-        """Enable materialization for an endpoint version.
-
-        If version is not specified, uses the current version.
-        Each version gets its own saved_query with naming: {endpoint_name}_v{version}
-        """
-        try:
-            self._enable_materialization_inner(endpoint, data_freshness_seconds, request, version, bucket_overrides)
-            ENDPOINT_MATERIALIZATION_EVENT_TOTAL.labels(action="enable", status="success").inc()
-            target_version = version or endpoint.get_version()
-            if target_version and target_version.saved_query:
-                log_activity(
-                    organization_id=self.organization.id,
-                    team_id=self.team.id,
-                    user=cast(User, request.user),
-                    was_impersonated=is_impersonated_session(request),
-                    item_id=str(target_version.saved_query.id),
-                    scope="DataWarehouseSavedQuery",
-                    activity="materialization_enabled",
-                    detail=Detail(
-                        name=target_version.saved_query.name,
-                        context=EndpointContext(version=target_version.version),
-                    ),
-                )
-        except ValidationError:
-            ENDPOINT_MATERIALIZATION_EVENT_TOTAL.labels(action="enable", status="validation_error").inc()
-            raise
-        except Exception:
-            ENDPOINT_MATERIALIZATION_EVENT_TOTAL.labels(action="enable", status="error").inc()
-            raise ValidationError("Failed to enable materialization.")
-
-    def _enable_materialization_inner(
-        self,
-        endpoint: Endpoint,
-        data_freshness_seconds: int,
-        request: Request,
-        version: EndpointVersion | None = None,
-        bucket_overrides: dict[str, str] | None = None,
-    ) -> None:
-        version = version or endpoint.get_version()
-
-        can_mat, reason = version.can_materialize()
-        if not can_mat:
-            raise ValidationError(f"Cannot materialize endpoint. Reason: {reason}")
-
-        # Per-version naming allows independent materialization for each version
-        saved_query_name = f"{endpoint.name}_v{version.version}"
-        saved_query = DataWarehouseSavedQuery.objects.filter(
-            name=saved_query_name, team=self.team, deleted=False
-        ).first()
-        if saved_query is None:
-            saved_query = DataWarehouseSavedQuery(
-                name=saved_query_name,
-                team=self.team,
-                origin=DataWarehouseSavedQuery.Origin.ENDPOINT,
-            )
-
-        hogql_query = build_endpoint_hogql(version.query, self.team, bucket_overrides=bucket_overrides)
-
-        saved_query.query = hogql_query
-        saved_query.external_tables = saved_query.s3_tables
-        saved_query.is_materialized = True
-        saved_query.sync_frequency_interval = sync_frequency_to_sync_frequency_interval(
-            DATA_FRESHNESS_BUCKETS[data_freshness_seconds]
-        )
-
-        saved_query.save()
-
-        version.saved_query = saved_query
-        version.bucket_overrides = bucket_overrides
-        version.save(update_fields=["saved_query", "bucket_overrides", "updated_at"])
-
-        saved_query.schedule_materialization()
-
-        try:
-            sync_saved_query_to_dag(saved_query)
-        except Exception as e:
-            logger.exception(
-                "Failed to sync endpoint node to DAG",
-                endpoint_name=endpoint.name,
-                saved_query_id=version.saved_query.id if version and version.saved_query else None,
-            )
-            capture_exception(
-                e,
-                {
-                    "product": Product.ENDPOINTS,
-                    "team_id": self.team_id,
-                    "endpoint_name": endpoint.name,
-                    "saved_query_id": version.saved_query.id if version and version.saved_query else None,
-                },
-            )
-
-    def _disable_materialization(
-        self, endpoint: Endpoint, request: Request, version: EndpointVersion | None = None
-    ) -> None:
-        """Disable materialization for an endpoint version.
-
-        If version is not specified, uses the current version.
-        """
-        version = version or endpoint.get_version()
-        if version:
-            if version.saved_query:
-                saved_query_id = str(version.saved_query.id)
-                saved_query_name = version.saved_query.name
-                try:
-                    delete_node_from_dag(version.saved_query)
-                except Exception as e:
-                    logger.exception(
-                        "Failed to remove endpoint node from DAG",
-                        endpoint_name=endpoint.name,
-                        saved_query_id=version.saved_query.id if version and version.saved_query else None,
-                    )
-                    capture_exception(
-                        e,
-                        {
-                            "product": Product.ENDPOINTS,
-                            "team_id": self.team_id,
-                            "endpoint_name": endpoint.name,
-                            "saved_query_id": version.saved_query.id if version and version.saved_query else None,
-                        },
-                    )
-                try:
-                    version.disable_materialization()
-                except Exception:
-                    ENDPOINT_MATERIALIZATION_EVENT_TOTAL.labels(action="disable", status="error").inc()
-                    raise
-                ENDPOINT_MATERIALIZATION_EVENT_TOTAL.labels(action="disable", status="success").inc()
-                log_activity(
-                    organization_id=self.organization.id,
-                    team_id=self.team.id,
-                    user=cast(User, request.user),
-                    was_impersonated=is_impersonated_session(request),
-                    item_id=saved_query_id,
-                    scope="DataWarehouseSavedQuery",
-                    activity="materialization_disabled",
-                    detail=Detail(
-                        name=saved_query_name,
-                        context=EndpointContext(version=version.version),
-                    ),
-                )
-        clear_endpoint_materialization_cache(self.team_id, endpoint.name)
+        return Response(result)
 
     def destroy(self, request: Request, name=None, *args, **kwargs) -> Response:
         """Delete an endpoint and clean up materialized query."""
         endpoint = self._get_endpoint_with_object_access(name)
-        endpoint_id = str(endpoint.id)
-        endpoint_name = endpoint.name
-
-        for version in endpoint.versions.filter(saved_query__isnull=False):
-            try:
-                if version.saved_query:
-                    delete_node_from_dag(version.saved_query)
-            except Exception as e:
-                logger.exception(
-                    "Failed to remove endpoint node from DAG on destroy",
-                    endpoint_name=endpoint.name,
-                    saved_query_id=version.saved_query.id if version.saved_query else None,
-                )
-                capture_exception(
-                    e,
-                    {
-                        "product": Product.ENDPOINTS,
-                        "team_id": self.team_id,
-                        "endpoint_name": endpoint.name,
-                        "saved_query_id": version.saved_query.id if version and version.saved_query else None,
-                    },
-                )
-
-        endpoint.soft_delete()
-        clear_endpoint_materialization_cache(self.team_id, endpoint.name)
-        log_activity(
-            organization_id=self.organization.id,
-            team_id=self.team.id,
-            user=cast(User, request.user),
-            was_impersonated=is_impersonated_session(request),
-            item_id=endpoint_id,
-            scope="Endpoint",
-            activity="deleted",
-            detail=Detail(name=endpoint_name),
-        )
+        EndpointCrudService(self.team, request).destroy(endpoint)
         return Response(status=status.HTTP_204_NO_CONTENT)
 
-    def _should_use_materialized_table(
-        self, endpoint: Endpoint, data: EndpointRunRequest, version: EndpointVersion | None = None
-    ) -> bool:
-        """
-        Decide whether to use materialized table or inline execution.
-
-        Returns False if:
-        - Not materialized
-        - Materialization incomplete/failed
-        - Materialized data is stale (older than sync frequency)
-        - User overrides present (variables, query)
-        - 'direct' mode requested (explicitly bypass materialization)
-        """
-        version = version or endpoint.get_version()
-        if not version.is_materialized or not version.saved_query:
-            return False
-
-        saved_query = version.saved_query
-        if saved_query.status not in ["Completed"]:
-            return False
-
-        if not saved_query.table:
-            return False
-
-        # Check if materialized data is stale
-        if saved_query.last_run_at and saved_query.sync_frequency_interval:
-            next_refresh_due = saved_query.last_run_at + saved_query.sync_frequency_interval
-            if timezone.now() >= next_refresh_due:
-                return False
-
-        # 'direct' mode explicitly bypasses materialization to run the original query
-        if data.refresh == EndpointRefreshMode.DIRECT:
-            return False
-
-        # Check if variables are valid for materialized execution
-        if data.variables:
-            query = version.query
-            query_kind = query.get("kind")
-
-            if query_kind == "HogQLQuery":
-                # HogQL: check if request variables are a subset of materialized variables
-                materialized_vars = self._get_materialized_variables(version)
-                if not materialized_vars:
-                    return False
-
-                materialized_codes = {v.code_name for v in materialized_vars}
-                request_var_codes = set(data.variables.keys())
-                if not request_var_codes.issubset(materialized_codes):
-                    return False
-            else:
-                # Materialized insight: only breakdown properties allowed
-                breakdown_filter = query.get("breakdownFilter") or {}
-                allowed_props = set(_get_breakdown_properties(breakdown_filter))
-                if not allowed_props:
-                    return False
-
-                request_var_codes = set(data.variables.keys())
-                if not request_var_codes.issubset(allowed_props):
-                    return False
-
-        return True
-
-    def _get_materialized_variables(self, version: EndpointVersion) -> builtins.list:
-        """Return the materializable variable infos for an endpoint query."""
-        query = version.query
-        if not query or not query.get("variables"):
-            return []
-
-        try:
-            can_materialize, _, variable_infos = analyze_variables_for_materialization(
-                query, bucket_overrides=version.bucket_overrides
-            )
-            return variable_infos if can_materialize else []
-        except Exception:
-            logger.debug("Failed to analyze variables for materialization", exc_info=True)
-            return []
-
-    def _parse_original_hogql_query(self, query: dict) -> tuple[builtins.list | None, int | None, bool]:
-        """Parse the original HogQL query and extract SELECT columns, LIMIT, and GROUP BY presence.
-
-        Returns (materialized_columns, limit, has_group_by) where columns/limit may be None.
-        materialized_columns is a list of MaterializedColumn with re-aggregation metadata.
-        """
-        query_str = query.get("query")
-        if not query_str:
-            return None, None, False
-
-        try:
-            parsed = parse_select(query_str)
-            if isinstance(parsed, ast.SelectQuery):
-                columns = (
-                    transform_select_for_materialized_table(list(parsed.select), self.team) if parsed.select else None
-                )
-                limit = parsed.limit.value if isinstance(parsed.limit, ast.Constant) else None
-                has_group_by = bool(parsed.group_by)
-                return columns, limit, has_group_by
-        except Exception:
-            logger.debug("Failed to parse original HogQL query", exc_info=True)
-
-        return None, None, False
-
-    # Query types that support user-configurable breakdown filtering
-    BREAKDOWN_SUPPORTED_QUERY_TYPES = {"TrendsQuery", "RetentionQuery"}
-
-    def _get_allowed_variables(self, query: dict, is_materialized: bool) -> set[str]:
-        """Get the set of allowed variable names for this endpoint."""
-        query_kind = query.get("kind")
-
-        if query_kind == "HogQLQuery":
-            # HogQL: allowed variables are code_names from query["variables"]
-            variables = query.get("variables", {})
-            return {v.get("code_name") for v in variables.values() if v.get("code_name")} if variables else set()
-
-        # Insight queries
-        allowed: set[str] = set()
-
-        # Only allow breakdown properties for query types that support it
-        if query_kind in self.BREAKDOWN_SUPPORTED_QUERY_TYPES:
-            breakdown_filter = query.get("breakdownFilter") or {}
-            allowed.update(_get_breakdown_properties(breakdown_filter))
-
-        if not is_materialized:
-            # Non-materialized also allows date_from/date_to via filters_override
-            allowed.update({"date_from", "date_to"})
-
-        return allowed
-
-    def _get_required_variables_for_materialized(self, query: dict, version: EndpointVersion) -> set[str]:
-        """Get the required variable names for a materialized endpoint.
-
-        SECURITY: This prevents data leakage by ensuring that materialized endpoints
-        with variables cannot be called without providing all variable values.
-        """
-        query_kind = query.get("kind")
-
-        if query_kind == "HogQLQuery":
-            materialized_vars = self._get_materialized_variables(version)
-            return {v.code_name for v in materialized_vars}
-
-        # Insight queries: all breakdown properties are required
-        if query_kind in self.BREAKDOWN_SUPPORTED_QUERY_TYPES:
-            breakdown_filter = query.get("breakdownFilter") or {}
-            return set(_get_breakdown_properties(breakdown_filter))
-
-        return set()
-
-    def _apply_where_filter(
-        self,
-        select_query: ast.SelectQuery,
-        column: str,
-        value: str,
-        op: ast.CompareOperationOp = ast.CompareOperationOp.Eq,
-        value_wrapper_fns: builtins.list[str] | None = None,
-        bucket_fn: str | None = None,
-    ) -> None:
-        """Add a comparison filter to WHERE clause.
-
-        When bucket_fn is set (range variables), wrap the filter value with the
-        bucket function so the comparison matches the bucketed column values.
-        """
-        right_expr: ast.Expr = ast.Constant(value=value)
-        for fn in reversed(value_wrapper_fns or []):
-            right_expr = ast.Call(name=fn, args=[right_expr])
-        if bucket_fn:
-            right_expr = ast.Call(name=bucket_fn, args=[ast.Call(name="toDateTime", args=[right_expr])])
-        condition = ast.CompareOperation(
-            left=ast.Field(chain=[column]),
-            op=op,
-            right=right_expr,
-        )
-        _add_where_condition(select_query, condition)
-
-    def _build_breakdown_filter_condition(
-        self, query_kind: str | None, value: str, array_index: int = 0
-    ) -> ast.Expr | None:
-        """Build the appropriate WHERE condition for breakdown filtering based on query type.
-
-        Different insight types store breakdowns in different columns:
-        - TrendsQuery, RetentionQuery: `breakdown_value` Array column
-        - LifecycleQuery: No breakdown support
-
-        array_index controls how to access the breakdown column:
-        - 0: single breakdown — use has() on the full array
-        - 1+: multiple breakdowns — use equality on breakdown_value[N]
-        """
-        if query_kind not in ("TrendsQuery", "RetentionQuery"):
-            logger.warning(
-                "Query type does not support breakdown filtering",
-                query_kind=query_kind,
-            )
-            return None
-
-        if array_index > 0:
-            return ast.CompareOperation(
-                left=ast.ArrayAccess(
-                    array=ast.Field(chain=["breakdown_value"]),
-                    property=ast.Constant(value=array_index),
-                ),
-                op=ast.CompareOperationOp.Eq,
-                right=ast.Constant(value=value),
-            )
-
-        return ast.Call(
-            name="has",
-            args=[ast.Field(chain=["breakdown_value"]), ast.Constant(value=value)],
-        )
-
-    def _execute_query_and_respond(
-        self,
-        query_request_data: dict,
-        client_query_id: str | None,
-        request: Request,
-        variables_override: builtins.list[HogQLVariable] | None = None,
-        cache_age_seconds: int | None = None,
-        extra_result_fields: dict | None = None,
-        debug: bool = False,
-        headers: dict[str, str] | None = None,
-        pagination: EndpointPagination | None = None,
-    ) -> Response:
-        """Shared query execution logic."""
-        merged_data = self.get_model(query_request_data, QueryRequest)
-
-        logger.debug(merged_data)
-        query, client_query_id, execution_mode = _process_query_request(
-            merged_data, self.team, client_query_id, request.user
-        )
-        self._tag_client_query_id(client_query_id)
-        endpoint_feature = (
-            Feature.ENDPOINT_EXECUTION if get_query_tag_value("access_method") else Feature.ENDPOINT_PLAYGROUND
-        )
-        tag_queries(product=Product.ENDPOINTS, feature=endpoint_feature)
-
-        if execution_mode not in BLOCKING_EXECUTION_MODES:
-            raise ValidationError({"refresh": f"Only sync modes are supported, got: {execution_mode}"})
-
-        result = process_query_model(
-            self.team,
-            query,
-            variables_override=variables_override,
-            execution_mode=execution_mode,
-            query_id=client_query_id,
-            user=cast(User, request.user),
-            is_query_service=is_api_key_access_method(get_query_tag_value("access_method")),
-            cache_age_seconds=cache_age_seconds,
-            analytics_props=get_request_analytics_properties(request),
-        )
-
-        if isinstance(result, BaseModel):
-            result = result.model_dump(by_alias=True)
-
-        if isinstance(result, dict) and extra_result_fields:
-            result.update(extra_result_fields)
-
-        if not debug:
-            debug_fields_to_remove = [
-                "calculation_trigger",
-                "cache_key",
-                "explain",
-                "modifiers",
-                "resolved_date_range",
-                "timings",
-                "hogql",
-            ]
-
-            for field in debug_fields_to_remove:
-                result.pop(field, None)
-
-        if pagination and "results" in result:
-            pagination.process_results(result)
-        elif "results" in result:
-            result["hasMore"] = False
-
-        _clean_breakdown_sentinels(result)
-
-        if "results" in result:
-            results_value = result.pop("results")
-            result = {"results": results_value, **result}
-
-        response_status = (
-            status.HTTP_202_ACCEPTED
-            if result.get("query_status") and result["query_status"].get("complete") is False
-            else status.HTTP_200_OK
-        )
-        return Response(result, status=response_status, headers=headers)
-
-    def _is_cache_stale(self, result: Response, saved_query) -> bool:
-        """Check if cached result is older than the materialization."""
-        if not isinstance(result.data, dict) or not result.data.get("is_cached"):
-            return False
-
-        last_refresh = result.data.get("last_refresh")
-        if not last_refresh or not saved_query.last_run_at:
-            return False
-
-        if isinstance(last_refresh, str):
-            last_refresh = isoparse(last_refresh)
-
-        return last_refresh < saved_query.last_run_at
-
-    def _build_materialized_select_query(
-        self,
-        table_name: str,
-        version: EndpointVersion,
-        variable_infos: builtins.list | None = None,
-    ) -> tuple[ast.SelectQuery, int | None]:
-        """Build the base SELECT query against a materialized table.
-
-        Wraps aggregate columns with their reaggregate_fn (e.g. sum("count()"))
-        when needed to preserve SQL semantics. This happens in two cases:
-
-        1. Range variables (bucket_fn != None): multiple materialized rows must
-           be collapsed back into one aggregate per group.
-        2. No GROUP BY + all-aggregate SELECT: without wrapping, an empty WHERE
-           match returns 0 rows instead of the expected 1-row identity result
-           (count→0, sum→0, min/max→type default). Wrapping restores the implicit
-           single-group aggregate that SQL guarantees.
-
-        Non-aggregate columns (GROUP BY dimensions) are passed through as-is and
-        added to the GROUP BY clause.
-
-        Returns (select_query, original_limit) — caller is responsible for pagination.
-        Used by both execution and preview paths.
-        """
-        query = version.query
-        query_kind = query.get("kind")
-
-        select_columns: list[ast.Expr] = [ast.Field(chain=["*"])]
-        group_by_columns: list[ast.Expr] = []
-        original_limit: int | None = None
-
-        if query_kind == "HogQLQuery":
-            original_select, original_limit, has_group_by = self._parse_original_hogql_query(query)
-            mat_vars = variable_infos if variable_infos is not None else self._get_materialized_variables(version)
-            if mat_vars and original_select:
-                has_range_vars = any(v.bucket_fn is not None for v in mat_vars)
-                all_aggregates = all(col.is_aggregate and col.reaggregate_fn for col in original_select)
-                needs_reaggregation = has_range_vars or (not has_group_by and all_aggregates)
-
-                if needs_reaggregation:
-                    reagg_select: list[ast.Expr] = []
-                    for col in original_select:
-                        if col.is_aggregate and col.reaggregate_fn:
-                            reagg_select.append(ast.Call(name=col.reaggregate_fn, args=[col.expr]))
-                        else:
-                            reagg_select.append(col.expr)
-                            if not col.is_aggregate:
-                                group_by_columns.append(col.expr)
-                    select_columns = reagg_select
-                else:
-                    select_columns = [col.expr for col in original_select]
-
-        select_query = ast.SelectQuery(
-            select=select_columns,
-            select_from=ast.JoinExpr(table=ast.Field(chain=[table_name])),
-        )
-
-        if group_by_columns:
-            select_query.group_by = [CloningVisitor().visit(c) for c in group_by_columns]
-
-        if original_limit is not None:
-            select_query.limit = ast.Constant(value=original_limit)
-
-        return select_query, original_limit
-
-    def _execute_materialized_endpoint(
-        self,
-        endpoint: Endpoint,
-        data: EndpointRunRequest,
-        request: Request,
-        version: EndpointVersion | None = None,
-        debug: bool = False,
-        limit: int | None = None,
-        offset: int | None = None,
-    ) -> Response:
-        """Execute against a materialized table in S3."""
-        materialized_hogql_query = None
-        query_kind = None
-        try:
-            version = version or endpoint.get_version()
-            if not version.saved_query:
-                raise ValidationError("No materialized query found for this endpoint")
-            saved_query = version.saved_query
-
-            query = version.query
-            query_kind = query.get("kind")
-
-            select_query, original_limit = self._build_materialized_select_query(
-                table_name=saved_query.name,
-                version=version,
-            )
-
-            pagination: EndpointPagination | None = None
-
-            # Only paginate flat-row HogQL results. Insight types get transformed
-            # into nested structures where flat-row LIMIT/OFFSET is meaningless.
-            if limit is not None and query_kind == "HogQLQuery":
-                pagination = EndpointPagination(limit=limit, offset=offset or 0, ceiling=original_limit)
-                pagination.apply_to(select_query)
-
-            deprecation_headers: dict[str, str] | None = None
-
-            # For insight endpoints: filters_override takes precedence over variables (backwards compat)
-            if query_kind != "HogQLQuery" and data.filters_override is not None:
-                deprecation_headers = {
-                    "X-PostHog-Warn": "filters_override is deprecated. Use variables instead: https://posthog.com/docs/api/endpoints"
-                }
-                # Extract breakdown filter from properties
-                if data.filters_override.properties:
-                    for prop in data.filters_override.properties:
-                        if hasattr(prop, "key") and hasattr(prop, "value") and prop.value is not None:
-                            # Convert value to string for breakdown filter
-                            value = prop.value[0] if isinstance(prop.value, list) else prop.value
-                            condition = self._build_breakdown_filter_condition(query_kind, str(value))
-                            if condition:
-                                _add_where_condition(select_query, condition)
-                            # filters_override is deprecated — only the first property is used
-                            break
-            elif data.variables:
-                if query_kind == "HogQLQuery":
-                    # HogQL: filter by all materialized variable columns
-                    materialized_vars = self._get_materialized_variables(version)
-                    for mat_var in materialized_vars:
-                        var_value = data.variables.get(mat_var.code_name)
-                        if var_value is not None:
-                            self._apply_where_filter(
-                                select_query,
-                                mat_var.code_name,
-                                var_value,
-                                op=mat_var.operator,
-                                value_wrapper_fns=mat_var.value_wrapper_fns,
-                                bucket_fn=mat_var.bucket_fn,
-                            )
-                else:
-                    # Insight: filter by breakdown property names
-                    breakdown_filter = query.get("breakdownFilter") or {}
-                    index_mapping = _get_breakdown_column_indices(breakdown_filter)
-
-                    for breakdown_prop, array_index in index_mapping.items():
-                        if breakdown_prop in data.variables:
-                            condition = self._build_breakdown_filter_condition(
-                                query_kind, data.variables[breakdown_prop], array_index=array_index
-                            )
-                            if condition:
-                                _add_where_condition(select_query, condition)
-
-            materialized_hogql_query = HogQLQuery(
-                query=select_query.to_hogql(),
-                modifiers=HogQLQueryModifiers(useMaterializedViews=True),
-            )
-
-            refresh_type = _endpoint_refresh_mode_to_refresh_type(data.refresh)
-
-            query_request_data = {
-                "client_query_id": data.client_query_id,
-                "name": f"{endpoint.name}_materialized",
-                "refresh": refresh_type,
-                "query": materialized_hogql_query.model_dump(),
-            }
-
-            extra_fields = {
-                "endpoint_materialized": True,
-                "endpoint_materialized_at": saved_query.last_run_at.isoformat() if saved_query.last_run_at else None,
-            }
-            tag_queries(
-                workload=Workload.ENDPOINTS,
-                warehouse_query=True,
-                endpoint_version=version.version if version else None,
-            )
-
-            # Compute dynamic cache TTL: time remaining until data_freshness window expires
-            cache_ttl = None
-            if saved_query.last_run_at and version.data_freshness_seconds:
-                remaining = (
-                    saved_query.last_run_at + timedelta(seconds=version.data_freshness_seconds) - timezone.now()
-                ).total_seconds()
-                if remaining <= 0:
-                    logger.warning(
-                        "endpoint_materialization_behind_sla",
-                        endpoint_name=endpoint.name,
-                        team_id=self.team_id,
-                        data_freshness_seconds=version.data_freshness_seconds,
-                        last_run_at=saved_query.last_run_at.isoformat(),
-                        remaining_seconds=remaining,
-                    )
-                    tag_queries(endpoint_materialization_behind=True)
-                cache_ttl = max(1, int(remaining))  # at least 1 second to enable caching
-
-            result = self._execute_query_and_respond(
-                query_request_data,
-                data.client_query_id,
-                request,
-                cache_age_seconds=cache_ttl,
-                extra_result_fields=extra_fields,
-                debug=debug,
-                headers=deprecation_headers,
-                pagination=pagination,
-            )
-
-            if self._is_cache_stale(result, saved_query):
-                query_request_data["refresh"] = RefreshType.FORCE_BLOCKING
-                result = self._execute_query_and_respond(
-                    query_request_data,
-                    data.client_query_id,
-                    request,
-                    extra_result_fields=extra_fields,
-                    debug=debug,
-                    headers=deprecation_headers,
-                    pagination=pagination,
-                )
-
-            INSIGHT_TRANSFORM_TYPES = {"TrendsQuery", "LifecycleQuery", "RetentionQuery"}
-            if query_kind in INSIGHT_TRANSFORM_TYPES:
-                try:
-                    transform_materialized_insight_response(
-                        result.data,
-                        query,
-                        self.team,
-                        now=saved_query.last_run_at,
-                    )
-                except MaterializedSeriesMismatchError:
-                    # Series drift: query was likely edited after materialization. Trigger a refresh
-                    # so the next request succeeds, but fail this one loudly rather than
-                    # returning wrong-labeled data.
-                    logger.warning(
-                        "Materialized endpoint series mismatch, triggering re-materialization",
-                        endpoint_name=endpoint.name,
-                        saved_query_id=saved_query.id,
-                    )
-                    trigger_saved_query_schedule(saved_query)
-                    raise
-
-            # Freshness relative to the configured target: >1.0 means behind SLA.
-            # Absolute age is meaningless across endpoints with different frequencies.
-            if saved_query.last_run_at and version and version.data_freshness_seconds:
-                age_seconds = max((timezone.now() - saved_query.last_run_at).total_seconds(), 0.0)
-                ENDPOINT_MATERIALIZED_FRESHNESS_RATIO.observe(age_seconds / version.data_freshness_seconds)
-
-            return result
-        except Exception as e:
-            logger.exception(
-                "Materialized endpoint execution failed",
-                endpoint_name=endpoint.name,
-                saved_query_id=saved_query.id if saved_query else None,
-                saved_query_status=saved_query.status if saved_query else None,
-            )
-            capture_exception(
-                e,
-                {
-                    "product": Product.ENDPOINTS,
-                    "team_id": self.team_id,
-                    "endpoint_name": endpoint.name,
-                    "materialized": True,
-                    "saved_query_id": saved_query.id if saved_query else None,
-                },
-            )
-            _emit_endpoint_failure_signal(
-                self.team,
-                endpoint,
-                e,
-                materialized=True,
-                version=version.version if version else None,
-                saved_query_id=saved_query.id if saved_query else None,
-                query_kind=query_kind,
-                executed_sql=materialized_hogql_query.query if materialized_hogql_query else None,
-                saved_query_status=saved_query.status if saved_query else None,
-                saved_query_last_run_at=(
-                    saved_query.last_run_at.isoformat() if saved_query and saved_query.last_run_at else None
-                ),
-                saved_query_columns=saved_query.columns if saved_query else None,
-                endpoint_columns=version.columns if version else None,
-            )
-            raise
-
-    @staticmethod
-    def _parse_int_param(
-        body_value: int | None, query_param: str | None, name: str, min_value: int | None = None
-    ) -> int | None:
-        """Parse an integer from the request body or query params. Raises `_RunRejected` if invalid."""
-        value = body_value
-        if value is None and query_param is not None:
-            try:
-                value = int(query_param)
-                if min_value is not None and value < min_value:
-                    raise ValueError()
-            except (ValueError, TypeError):
-                raise _RunRejected(
-                    Response({"error": f"Invalid {name} parameter: {query_param}"}, status=status.HTTP_400_BAD_REQUEST)
-                )
-        elif value is not None and min_value is not None and value < min_value:
-            raise _RunRejected(
-                Response({"error": f"Invalid {name} parameter: {value}"}, status=status.HTTP_400_BAD_REQUEST)
-            )
-        return value
-
-    def _apply_pagination_to_query(self, query: dict, limit: int, offset: int) -> tuple[dict, EndpointPagination]:
-        """Apply pagination to a HogQL query. Returns (modified_query, pagination).
-
-        Parses the HogQL AST, applies LIMIT/OFFSET, and reprints using a
-        placeholder-preserving printer so unresolved {variables.*} survive.
-        """
-        query_kind = query.get("kind")
-
-        if query_kind == "HogQLQuery":
-            query_sql = query.get("query", "")
-            parsed = parse_select(query_sql)
-            if not isinstance(parsed, ast.SelectQuery):
-                raise ValidationError(
-                    "Pagination is not supported for UNION queries. Wrap the UNION in a subquery: SELECT * FROM (... UNION ALL ...) LIMIT n"
-                )
-
-            ceiling = parsed.limit.value if isinstance(parsed.limit, ast.Constant) else None
-            pagination = EndpointPagination(limit=limit, offset=offset, ceiling=ceiling)
-
-            pagination.apply_to(parsed)
-
-            ctx = HogQLContext(enable_select_queries=True, limit_top_select=False)
-            query = query.copy()
-            query["query"] = _PlaceholderPreservingPrinter(context=ctx).visit(parsed)
-            return query, pagination
-
-        raise ValidationError({"limit": f"Limit/offset parameters are only supported for HogQLQuery, not {query_kind}"})
-
-    def _parse_variables(
-        self, query: dict[str, dict], variables: dict[str, str]
-    ) -> builtins.list[HogQLVariable] | None:
-        query_variables = query.get("variables", None)
-        if not query_variables:
-            return None
-
-        variables_override = []
-        for request_variable_code_name, request_variable_value in variables.items():
-            variable_id = None
-            for query_variable_id, query_variable_value in query_variables.items():
-                if query_variable_value.get("code_name", None) == request_variable_code_name:
-                    variable_id = query_variable_id
-
-            if variable_id is None:
-                raise ValidationError({"variables": f"Variable '{request_variable_code_name}' not found in query"})
-
-            variables_override.append(
-                HogQLVariable(
-                    variableId=variable_id,
-                    code_name=request_variable_code_name,
-                    value=request_variable_value,
-                    isNull=True if request_variable_value is None else None,
-                )
-            )
-        return variables_override
-
-    def _variables_to_filters(
-        self,
-        variables: dict[str, str],
-        breakdown_infos: builtins.list[tuple[str, str]] | None = None,
-    ):
-        """Convert insight magic variables to DashboardFilter.
-
-        Args:
-            variables: Dict of variable name -> value from the request
-            breakdown_infos: List of (property_name, property_type) for breakdown filtering
-        """
-
-        date_from = variables.get("date_from")
-        date_to = variables.get("date_to")
-
-        properties: list[dict] = []
-        for prop_name, prop_type in breakdown_infos or []:
-            if prop_name not in variables:
-                continue
-            value = variables[prop_name]
-            # Any failure to build a filter for a breakdown variable must error the
-            # request — silently skipping would return unfiltered data, which is a
-            # data-leak (same principle that gates materialized endpoints on having
-            # all variables present).
-            try:
-                if prop_type == "hogql":
-                    # HogQLPropertyFilter.key is a full HogQL predicate and doesn't accept
-                    # `operator` (HogQLPropertyFilter has extra="forbid"), so build a
-                    # `<expr> = <value>` (or `isNull(<expr>)`) predicate to filter with.
-                    key_expr = parse_expr(prop_name)
-                    predicate_expr: ast.Expr
-                    if value == "":
-                        predicate_expr = ast.Call(name="isNull", args=[key_expr])
-                    else:
-                        predicate_expr = ast.CompareOperation(
-                            left=key_expr,
-                            op=ast.CompareOperationOp.Eq,
-                            right=ast.Constant(value=value),
-                        )
-                    properties.append({"key": predicate_expr.to_hogql(), "type": "hogql"})
-                    continue
-                if value == "":
-                    # Empty string means "null/unset" — match events where the property is missing.
-                    # This keeps inline execution consistent with the materialized path,
-                    # where null breakdown values are stored as '' in S3.
-                    properties.append(
-                        {
-                            "key": prop_name,
-                            "type": prop_type if prop_type else "event",
-                            "operator": PropertyOperator.IS_NOT_SET,
-                        }
-                    )
-                else:
-                    properties.append(
-                        {
-                            "key": prop_name,
-                            "value": value,
-                            "type": prop_type if prop_type else "event",
-                            "operator": PropertyOperator.EXACT,
-                        }
-                    )
-            except Exception as e:
-                capture_exception(
-                    e,
-                    {
-                        "product": Product.ENDPOINTS,
-                        "team_id": self.team_id,
-                        "endpoint_name": self.kwargs.get("name"),
-                        "prop_name": prop_name,
-                        "prop_type": prop_type,
-                    },
-                )
-                raise ValidationError(
-                    {"variables": f"Could not apply filter for breakdown variable '{prop_name}'"}
-                ) from e
-
-        if not date_from and not date_to and not properties:
-            return None
-
-        return DashboardFilter(date_from=date_from, date_to=date_to, properties=properties)
-
-    def _should_use_ducklake(self, endpoint: Endpoint, version: EndpointVersion | None) -> bool:
-        if version is None or version.query.get("kind") != "HogQLQuery":
-            return False
-
-        ff_result = posthoganalytics.feature_enabled(
-            "endpoints-ducklake-execution",
-            str(self.team.uuid),
-            groups={
-                "organization": str(self.team.organization_id),
-                "project": str(self.team.id),
-            },
-            group_properties={
-                "organization": {"id": str(self.team.organization_id)},
-                "project": {"id": str(self.team.id)},
-            },
-            only_evaluate_locally=True,
-            send_feature_flag_events=False,
-        )
-        logger.info("Ducklake FF evaluation", endpoint_name=endpoint.name, ff_result=ff_result)
-        if not ff_result:
-            return False
-
-        server = get_duckgres_server_for_organization(str(self.team.organization_id))
-        if server is None:
-            logger.info("Ducklake skip: no duckgres server", endpoint_name=endpoint.name, team_id=self.team_id)
-        return server is not None
-
-    def _execute_ducklake_endpoint(
-        self,
-        endpoint: Endpoint,
-        query: dict,
-        debug: bool = False,
-    ) -> Response:
-        from posthog.ducklake.client import execute_ducklake_query
-
-        try:
-            result = execute_ducklake_query(
-                self.team_id,
-                query=HogQLQuery(query=query["query"]),
-                organization_id=str(self.team.organization_id),
-            )
-            response_data: dict = {
-                "results": result.results,
-                "columns": result.columns,
-                "types": result.types,
-                "hasMore": False,
-                "backend": "ducklake",
-            }
-            if debug:
-                response_data["query"] = query.get("query")
-                response_data["hogql"] = result.hogql
-                response_data["ducklake_sql"] = result.sql
-            return Response(response_data, status=status.HTTP_200_OK)
-        except Exception as e:
-            logger.exception(
-                "DuckLake endpoint execution failed",
-                endpoint_name=endpoint.name,
-            )
-            capture_exception(
-                e,
-                {
-                    "product": Product.ENDPOINTS,
-                    "team_id": self.team_id,
-                    "ducklake": True,
-                    "endpoint_name": endpoint.name,
-                },
-            )
-            raise
-
-    def _execute_inline_endpoint(
-        self,
-        endpoint: Endpoint,
-        data: EndpointRunRequest,
-        request: Request,
-        query: dict,
-        version: EndpointVersion | None = None,
-        debug: bool = False,
-        limit: int | None = None,
-        offset: int | None = None,
-    ) -> Response:
-        """Execute query directly against ClickHouse."""
-        try:
-            query = prepare_insight_query_for_endpoint(query)
-
-            pagination: EndpointPagination | None = None
-            if limit is not None:
-                query, pagination = self._apply_pagination_to_query(query, limit, offset or 0)
-
-            refresh_type = _endpoint_refresh_mode_to_refresh_type(data.refresh)
-            query_kind = query.get("kind")
-
-            variables_override = None
-            filters_override = None
-            deprecation_headers: dict[str, str] | None = None
-
-            # For insight endpoints: filters_override takes precedence over variables (backwards compat)
-            if query_kind != "HogQLQuery" and data.filters_override is not None:
-                filters_override = data.filters_override
-                deprecation_headers = {
-                    "X-PostHog-Warn": "filters_override is deprecated. Use variables instead: https://posthog.com/docs/api/endpoints"
-                }
-            elif data.variables:
-                if query_kind == "HogQLQuery":
-                    variables_override = self._parse_variables(query, data.variables)
-                else:
-                    breakdown_filter = query.get("breakdownFilter") or {}
-                    breakdown_infos = _get_breakdown_infos(breakdown_filter)
-                    filters_override = self._variables_to_filters(data.variables, breakdown_infos=breakdown_infos)
-
-            query_request_data = {
-                "client_query_id": data.client_query_id,
-                "filters_override": filters_override.model_dump() if filters_override else None,
-                "name": endpoint.name,
-                "refresh": refresh_type,
-                "query": query,
-            }
-
-            cache_age = version.data_freshness_seconds if version else None
-            tag_queries(endpoint_version=version.version if version else None)
-
-            return self._execute_query_and_respond(
-                query_request_data,
-                data.client_query_id,
-                request,
-                variables_override=variables_override,
-                cache_age_seconds=cache_age,
-                debug=debug,
-                headers=deprecation_headers,
-                pagination=pagination,
-            )
-
-        except Exception as e:
-            self.handle_column_ch_error(e)
-            logger.exception(
-                "Inline endpoint execution failed",
-                endpoint_name=endpoint.name,
-            )
-            capture_exception(
-                e,
-                {
-                    "product": Product.ENDPOINTS,
-                    "team_id": self.team_id,
-                    "materialized": False,
-                    "endpoint_name": endpoint.name,
-                },
-            )
-            _emit_endpoint_failure_signal(
-                self.team,
-                endpoint,
-                e,
-                materialized=False,
-                version=version.version if version else None,
-                query_kind=query_kind,
-                executed_sql=query.get("query") if query_kind == "HogQLQuery" else None,
-                endpoint_columns=version.columns if version else None,
-            )
-            raise
-
-    def _log_rejected_run(self, endpoint: Endpoint, reason: str) -> None:
-        """Record a rejected run request as a failed execution so it surfaces in the endpoint logs.
-
-        Request validation runs before the execution block, so a rejected request would otherwise
-        never appear in the logs. Each rejection gets its own execution id.
-        """
-        log_endpoint_execution(
-            team_id=self.team_id,
-            endpoint_id=str(endpoint.id),
-            instance_id=str(uuid.uuid4()),
-            level="ERROR",
-            message=f"Endpoint execution failed · invalid request · {reason}",
-        )
-
-    def _resolve_version(
-        self, endpoint: Endpoint, version_number: int | None, name: str | None
-    ) -> EndpointVersion | None:
-        """Return the requested version, or None when no specific version was asked for and none exist.
-
-        Raises `_RunRejected` (404) only when a specific `version_number` was requested but not found —
-        matching the original behavior where an unspecified version falls through to the inline path.
-        """
-        try:
-            return endpoint.get_version(version_number)
-        except EndpointVersion.DoesNotExist:
-            if version_number is not None:
-                raise _RunRejected(
-                    Response(
-                        {
-                            "error": f"Version {version_number} not found for endpoint '{name}'",
-                            "current_version": endpoint.current_version,
-                        },
-                        status=status.HTTP_404_NOT_FOUND,
-                    )
-                )
-            return None
-
-    @staticmethod
-    def _format_validation_detail(detail: Any) -> str:
-        """Flatten a DRF ValidationError detail (str | list | dict, possibly nested) into one line."""
-        if isinstance(detail, dict):
-            return "; ".join(
-                f"{field}: {EndpointViewSet._format_validation_detail(messages)}" for field, messages in detail.items()
-            )
-        if isinstance(detail, list):
-            return "; ".join(EndpointViewSet._format_validation_detail(item) for item in detail)
-        return str(detail)
+    # ------------------------------------------------------------------
+    # Execution
+    # ------------------------------------------------------------------
 
     @staticmethod
     def _parse_run_request(request: Request) -> EndpointRunRequest:
-        """Parse and validate the run request body into a clean, field-level 400 on failure.
-
-        Raises a DRF ValidationError so the caller never sees pydantic's raw error dump. Rejection
-        logging is handled centrally in `run()`.
-        """
+        """Parse the run body into field-level errors instead of pydantic's raw error dump."""
         try:
             return EndpointRunRequest.model_validate(request.data)
         except PydanticValidationError as exc:
             field_errors = {
                 ".".join(str(part) for part in error["loc"]) or "body": error["msg"] for error in exc.errors()
             }
-            raise ValidationError(field_errors)
+            raise ValidationError(field_errors) from exc
+
+    @staticmethod
+    def _rejection_reason(response: Response) -> str:
+        data = response.data
+        if isinstance(data, dict):
+            return str(data.get("error", data))
+        return str(data)
 
     @extend_schema(
         request=EndpointRunRequest,
@@ -2355,284 +455,50 @@ class EndpointViewSet(
         """Execute endpoint with optional parameters."""
         endpoint = get_object_or_404(Endpoint, team=self.team, name=name, is_active=True, deleted=False)
         self._enforce_object_level_access(endpoint)
-
-        # All run-request validation lives in one place so every rejection — whether it raises a DRF
-        # ValidationError or returns a 4xx Response — is logged exactly once before we bail out.
+        service = EndpointExecutionService(self.team, request)
         try:
             data = self._parse_run_request(request)
-
-            report_user_action(
-                user=cast("User | SyntheticUser", request.user),
-                event="endpoint executed",
-                properties={
-                    "endpoint_id": str(endpoint.id),
-                    "endpoint_name": endpoint.name,
-                    "has_filters_override": bool(data.filters_override),
-                    "has_variables": bool(data.variables),
-                    "has_limit": data.limit is not None,
-                    "has_offset": data.offset is not None,
-                    "refresh_mode": data.refresh.value if data.refresh else None,
-                    "auth_method": "project_secret_api_key"
-                    if is_authenticated_via_project_secret_api_key(request)
-                    else "user",
-                },
-                team=self.team,
-                request=request,
-            )
-
-            version_number = self._parse_int_param(data.version, request.query_params.get("version"), "version")
-            limit = self._parse_int_param(data.limit, request.query_params.get("limit"), "limit", min_value=1)
-            offset = self._parse_int_param(data.offset, request.query_params.get("offset"), "offset", min_value=0)
-
-            if offset is not None and limit is None:
-                raise _RunRejected(
-                    Response({"error": "offset requires limit to be set"}, status=status.HTTP_400_BAD_REQUEST)
-                )
-
-            version_obj = self._resolve_version(endpoint, version_number, name)
-
-            self.validate_run_request(data, endpoint, version_obj)
-
-            if offset is not None and version_obj and version_obj.query.get("kind") != "HogQLQuery":
-                raise _RunRejected(
-                    Response(
-                        {"error": "offset is only supported for HogQL endpoints"}, status=status.HTTP_400_BAD_REQUEST
-                    )
-                )
         except ValidationError as exc:
-            self._log_rejected_run(endpoint, self._format_validation_detail(exc.detail))
+            service.log_rejected_run(endpoint, service.format_validation_detail(exc.detail))
             raise
-        except _RunRejected as rejected:
-            data_ = rejected.response.data
-            reason = data_.get("error", data_) if isinstance(data_, dict) else data_
-            self._log_rejected_run(endpoint, str(reason))
-            return rejected.response
 
-        # Check if we should use materialization for this version
-        use_materialized = self._should_use_materialized_table(endpoint, data, version_obj)
+        version_number, err = self._parse_int_param(data.version, request.query_params.get("version"), "version")
+        if err:
+            service.log_rejected_run(endpoint, self._rejection_reason(err))
+            return err
+        limit, err = self._parse_int_param(data.limit, request.query_params.get("limit"), "limit", min_value=1)
+        if err:
+            service.log_rejected_run(endpoint, self._rejection_reason(err))
+            return err
+        offset, err = self._parse_int_param(data.offset, request.query_params.get("offset"), "offset", min_value=0)
+        if err:
+            service.log_rejected_run(endpoint, self._rejection_reason(err))
+            return err
 
-        debug = data.debug or False
-        execution_type = "materialized" if use_materialized else "inline"
-        query_kind_metric = query_kind_label(version_obj.query if version_obj else None)
-        execution_status: str | None = None
-        execution_id = str(uuid.uuid4())
-        cache_outcome: str | None = None
-        result_row_count: int | None = None
-        error_label: str | None = None
-        _start_time = time.monotonic()
-        _duration = 0.0
+        if offset is not None and limit is None:
+            response = Response(
+                {"error": "offset requires limit to be set"},
+                status=status.HTTP_400_BAD_REQUEST,
+            )
+            service.log_rejected_run(endpoint, self._rejection_reason(response))
+            return response
 
+        version_obj = None
         try:
-            if use_materialized:
-                result = self._execute_materialized_endpoint(
-                    endpoint, data, request, version=version_obj, debug=debug, limit=limit, offset=offset
+            version_obj = endpoint.get_version(version_number)
+        except EndpointVersion.DoesNotExist:
+            if version_number is not None:
+                response = Response(
+                    {
+                        "error": f"Version {version_number} not found for endpoint '{name}'",
+                        "current_version": endpoint.current_version,
+                    },
+                    status=status.HTTP_404_NOT_FOUND,
                 )
-            else:
-                # Use version's query
-                if not version_obj:
-                    return Response(
-                        {"error": "No version found for this endpoint"},
-                        status=status.HTTP_404_NOT_FOUND,
-                    )
-                query_to_use = version_obj.query.copy()
+                service.log_rejected_run(endpoint, self._rejection_reason(response))
+                return response
 
-                use_ducklake = self._should_use_ducklake(endpoint, version_obj)
-                if use_ducklake:
-                    try:
-                        result = self._execute_ducklake_endpoint(endpoint, query_to_use, debug=debug)
-                        execution_type = "ducklake"
-                    except Exception:
-                        logger.warning(
-                            "DuckLake execution failed, falling back to inline",
-                            endpoint_name=endpoint.name,
-                        )
-                        ENDPOINT_DUCKLAKE_FALLBACK_TOTAL.inc()
-                        execution_type = "ducklake_fallback"
-                        result = self._execute_inline_endpoint(
-                            endpoint,
-                            data,
-                            request,
-                            query_to_use,
-                            version=version_obj,
-                            debug=debug,
-                            limit=limit,
-                            offset=offset,
-                        )
-                else:
-                    result = self._execute_inline_endpoint(
-                        endpoint,
-                        data,
-                        request,
-                        query_to_use,
-                        version=version_obj,
-                        debug=debug,
-                        limit=limit,
-                        offset=offset,
-                    )
-            execution_status = "success"
-        except (ExposedHogQLError, ExposedCHQueryError) as e:
-            execution_status = "user_error"
-            error_label = getattr(e, "code_name", None) or type(e).__name__
-            logger.exception(
-                "Endpoint execution failed",
-                endpoint_name=endpoint.name,
-                code_name=getattr(e, "code_name", None),
-            )
-            raise ValidationError("Query execution failed.", getattr(e, "code_name", None))
-        except HogVMException:
-            execution_status = "user_error"
-            error_label = "HogVMException"
-            logger.exception(
-                "Endpoint execution failed (HogVM)",
-                endpoint_name=endpoint.name,
-            )
-            raise ValidationError("Query execution failed: HogQL virtual machine error")
-        except ResolutionError:
-            execution_status = "user_error"
-            error_label = "ResolutionError"
-            logger.exception(
-                "Endpoint resolution failed",
-                endpoint_name=endpoint.name,
-            )
-            raise ValidationError("Query resolution failed: unable to resolve table or field references.")
-        except ConcurrencyLimitExceeded:
-            ENDPOINT_CONCURRENCY_REJECTED_TOTAL.labels(team_id=str(self.team_id)).inc()
-            raise Throttled(detail="Too many concurrent requests. Please try again later.")
-        except Exception as e:
-            execution_status = "error"
-            error_label = type(e).__name__
-            raise
-        finally:
-            if execution_status is not None:
-                _duration = time.monotonic() - _start_time
-                ENDPOINT_EXECUTION_DURATION_SECONDS.labels(
-                    execution_type=execution_type, query_kind=query_kind_metric
-                ).observe(_duration)
-                ENDPOINT_EXECUTION_TOTAL.labels(
-                    execution_type=execution_type, query_kind=query_kind_metric, status=execution_status
-                ).inc()
-            if execution_status in ("error", "user_error"):
-                log_endpoint_execution(
-                    team_id=self.team_id,
-                    endpoint_id=str(endpoint.id),
-                    instance_id=execution_id,
-                    level="ERROR",
-                    message=build_execution_message(
-                        succeeded=False,
-                        execution_type=execution_type,
-                        version=version_obj.version if version_obj else None,
-                        error=error_label,
-                    ),
-                )
-
-        try:
-            if isinstance(result.data, dict):
-                # DuckLake bypasses the query result cache entirely — don't claim hit/miss for it.
-                if execution_type != "ducklake":
-                    cache_outcome = "hit" if bool(result.data.get("is_cached")) else "miss"
-                    ENDPOINT_CACHE_RESULT_TOTAL.labels(
-                        execution_type=execution_type, query_kind=query_kind_metric, outcome=cache_outcome
-                    ).inc()
-
-                results_value = result.data.get("results")
-                if isinstance(results_value, list):
-                    result_row_count = len(results_value)
-                    if query_kind_metric == "hogql":
-                        ENDPOINT_HOGQL_RESULT_ROWS.labels(execution_type=execution_type).observe(result_row_count)
-        except Exception:
-            logger.debug("Failed to record endpoint result metrics", exc_info=True)
-
-        log_endpoint_execution(
-            team_id=self.team_id,
-            endpoint_id=str(endpoint.id),
-            instance_id=execution_id,
-            level="INFO",
-            message=build_execution_message(
-                succeeded=True,
-                execution_type=execution_type,
-                cache_outcome=cache_outcome,
-                duration_ms=round(_duration * 1000),
-                rows=result_row_count,
-                version=version_obj.version if version_obj else None,
-            ),
-        )
-
-        if is_api_key_access_method(get_query_tag_value("access_method")):
-            now = timezone.now()
-            throttle = timedelta(minutes=30)
-            if endpoint.last_executed_at is None or (now - endpoint.last_executed_at > throttle):
-                endpoint.last_executed_at = now
-                endpoint.save(update_fields=["last_executed_at"])
-            if version_obj is not None and (
-                version_obj.last_executed_at is None or (now - version_obj.last_executed_at > throttle)
-            ):
-                version_obj.last_executed_at = now
-                version_obj.save(update_fields=["last_executed_at"])
-
-        if isinstance(result.data, dict):
-            result.data["name"] = endpoint.name
-            result.data["execution_id"] = execution_id
-            if version_obj:
-                result.data["endpoint_version"] = version_obj.version
-                result.data["endpoint_version_created_at"] = version_obj.created_at.isoformat()
-
-        return result
-
-    def validate_run_request(
-        self, data: EndpointRunRequest, endpoint: Endpoint, version: EndpointVersion | None = None
-    ) -> None:
-        version = version or endpoint.get_version()
-
-        query = version.query
-        is_materialized = bool(version.is_materialized and version.saved_query)
-
-        if version and not version.is_active:
-            ENDPOINT_VALIDATION_ERROR_TOTAL.labels(reason="inactive_version").inc()
-            raise ValidationError(f"Version {version.version} is inactive and cannot be executed.")
-
-        query_kind = query.get("kind")
-
-        # Reject query_override (always)
-        if hasattr(data, "query_override") and data.query_override is not None:
-            raise ValidationError({"query_override": "Not allowed. Use variables instead."})
-
-        # Allow filters_override for insight endpoints (deprecated but backwards compatible)
-        # Reject for HogQL endpoints
-        if data.filters_override is not None:
-            if query_kind == "HogQLQuery":
-                raise ValidationError({"filters_override": "Not allowed for HogQL endpoints. Use variables instead."})
-
-        # Validate refresh mode
-        if data.refresh == EndpointRefreshMode.DIRECT and not is_materialized:
-            ENDPOINT_VALIDATION_ERROR_TOTAL.labels(reason="direct_refresh_not_materialized").inc()
-            raise ValidationError(
-                {
-                    "refresh": "'direct' refresh mode is only valid for materialized endpoints. "
-                    "Use 'cache' or 'force' instead, or enable materialization on this endpoint."
-                }
-            )
-
-        # Validate variables
-        if data.variables:
-            allowed_vars = self._get_allowed_variables(query, is_materialized)
-            unknown_vars = set(data.variables.keys()) - allowed_vars
-            if unknown_vars:
-                ENDPOINT_VALIDATION_ERROR_TOTAL.labels(reason="unknown_variable").inc()
-                raise ValidationError({"variables": f"Unknown variable(s): {', '.join(sorted(unknown_vars))}"})
-
-        # SECURITY: For materialized endpoints with required variables, ALL must be provided.
-        # Without this check, omitting variables would return ALL data instead of filtered data.
-        # Exception: filters_override is the deprecated way to provide filters for insight endpoints
-        if is_materialized and not (data.filters_override and data.filters_override.properties):
-            required_vars = self._get_required_variables_for_materialized(query, version)
-            if required_vars:
-                provided = set(data.variables.keys()) if data.variables else set()
-                missing = sorted(required_vars - provided)
-                if missing:
-                    ENDPOINT_VALIDATION_ERROR_TOTAL.labels(reason="missing_required_variable").inc()
-                    raise ValidationError(
-                        {"variables": f"Required variable(s) {', '.join(repr(v) for v in missing)} not provided"}
-                    )
+        return service.execute(endpoint, data, version_obj, limit=limit, offset=offset)
 
     @extend_schema(
         description=(
@@ -2674,19 +540,9 @@ class EndpointViewSet(
             capture_exception(e, {"product": Product.ENDPOINTS, "team_id": self.team_id})
             raise
 
-    def handle_column_ch_error(self, error):
-        if getattr(error, "message", None):
-            match = re.search(r"There's no column.*in table", error.message)
-            if match:
-                # TODO: remove once we support all column types
-                raise ValidationError(match.group(0) + ". Not all column types are fully supported yet.")
-        return
-
-    def _tag_client_query_id(self, query_id: str | None):
-        if query_id is None:
-            return
-
-        tag_queries(client_query_id=query_id)
+    # ------------------------------------------------------------------
+    # Versions + materialization
+    # ------------------------------------------------------------------
 
     @extend_schema(
         responses={200: EndpointVersionResponseSerializer(many=True)},
@@ -2719,16 +575,11 @@ class EndpointViewSet(
         """
         endpoint = self._get_endpoint_with_object_access(name)
 
-        version_number = self._parse_version_param(request)
-        if version_number is not None:
-            try:
-                version = endpoint.get_version(version_number)
-            except EndpointVersion.DoesNotExist:
-                return Response({"error": f"Version {version_number} not found"}, status=status.HTTP_404_NOT_FOUND)
-        else:
-            version = endpoint.get_version()
+        version = self._resolve_version_param(request, endpoint)
+        if isinstance(version, Response):
+            return version
 
-        return Response(self._build_materialization_info(version, endpoint_name=endpoint.name))
+        return Response(build_materialization_info(version, endpoint_name=endpoint.name))
 
     @validated_request(
         MaterializationPreviewRequestSerializer,
@@ -2751,124 +602,12 @@ class EndpointViewSet(
         else:
             version = endpoint.get_version()
 
-        can_mat, reason = version.can_materialize()
-        if not can_mat:
-            return _cant_materialize_response(reason)
-
         bucket_overrides = request.validated_data.get("bucket_overrides")
         if bucket_overrides:
-            _validate_bucket_overrides(bucket_overrides)
+            validate_bucket_overrides(bucket_overrides)
 
-        hogql_query = convert_insight_query_to_hogql(version.query, self.team)
-
-        range_pairs: list[dict] = []
-        aggregates: list[dict] = []
-        transformed_query_str: str | None = None
-        variable_infos: list = []
-
-        if version.query.get("variables"):
-            can_materialize_vars, var_reason, variable_infos = analyze_variables_for_materialization(
-                version.query, bucket_overrides=bucket_overrides
-            )
-
-            if not can_materialize_vars:
-                return _cant_materialize_response(var_reason)
-
-            if variable_infos:
-                transformed = transform_query_for_materialization(
-                    hogql_query, variable_infos, self.team, bucket_overrides=bucket_overrides
-                )
-                transformed_query_str = transformed.get("query")
-
-                # Extract range pairs grouped by column
-                seen_columns: dict[str, dict] = {}
-                for v in variable_infos:
-                    if v.bucket_fn is not None:
-                        col_key = ".".join(v.column_chain) if v.column_chain else v.column_expression
-                        if col_key not in seen_columns:
-                            seen_columns[col_key] = {
-                                "column": col_key,
-                                "variables": [],
-                                "bucket_fn": v.bucket_fn,
-                            }
-                        seen_columns[col_key]["variables"].append(v.code_name)
-                range_pairs = list(seen_columns.values())
-
-                query_str = hogql_query.get("query", "")
-                if query_str:
-                    try:
-                        parsed = parse_select(query_str)
-                    except Exception:
-                        logger.warning("materialization_preview: failed to parse HogQL for aggregate extraction")
-                    else:
-                        if isinstance(parsed, ast.SelectQuery) and parsed.select:
-                            for expr in parsed.select:
-                                agg_name = _extract_aggregate_name(expr)
-                                if agg_name:
-                                    reagg_info = get_reaggregation(agg_name)
-                                    reagg = reagg_info.reaggregate_fn if reagg_info else None
-                                    if isinstance(expr, ast.Alias):
-                                        label = expr.alias
-                                    else:
-                                        label = expr.to_hogql()
-                                    aggregates.append(
-                                        {
-                                            "expression": label,
-                                            "reaggregate_fn": reagg,
-                                        }
-                                    )
-        else:
-            # No variables — just show the converted query as-is
-            transformed_query_str = hogql_query.get("query")
-
-        # Build the execution query preview — what runs at request time against the materialized table
-        execution_query_str: str | None = None
-        display_execution_query_str: str | None = None
-        try:
-            saved_query_name = f"{endpoint.name}_v{version.version}"
-
-            def _build_exec_preview(table_name: str) -> ast.SelectQuery:
-                q, _ = self._build_materialized_select_query(
-                    table_name=table_name,
-                    version=version,
-                    variable_infos=variable_infos or None,
-                )
-                for v in variable_infos:
-                    self._apply_where_filter(
-                        q,
-                        v.code_name,
-                        f"{{variables.{v.code_name}}}",
-                        op=v.operator,
-                        value_wrapper_fns=v.value_wrapper_fns,
-                        bucket_fn=v.bucket_fn,
-                    )
-                return q
-
-            # Each call builds a fresh SelectQuery, so WHERE mutations don't leak between calls
-            execution_query_str = to_printed_hogql(_build_exec_preview(saved_query_name), team=self.team)
-
-            # Display variant uses the friendly endpoint name — printed without type resolution
-            # since the friendly name isn't a real table in the database
-            display_execution_query_str = print_prepared_ast(
-                node=_build_exec_preview(endpoint.name),
-                context=HogQLContext(team_id=self.team.pk, enable_select_queries=True),
-                dialect="hogql",
-                pretty=True,
-            )
-        except Exception:
-            logger.debug("Failed to build execution query preview", exc_info=True)
-
-        return Response(
-            {
-                "can_materialize": True,
-                "reason": None,
-                "transformed_query": transformed_query_str,
-                "execution_query": execution_query_str,
-                "display_execution_query": display_execution_query_str,
-                "range_pairs": range_pairs,
-                "aggregates": aggregates,
-            }
-        )
+        service = EndpointMaterializationService(self.team, request)
+        return Response(dataclasses.asdict(service.preview(endpoint, version, bucket_overrides)))
 
     @extend_schema(
         # url_path="openapi.json" would otherwise produce `..._openapi.json_retrieve` —
@@ -2907,85 +646,17 @@ class EndpointViewSet(
         spec = generate_openapi_spec(endpoint, self.team.id, request, version)
         return Response(spec, content_type="application/json")
 
+    # ------------------------------------------------------------------
+    # Tags
+    # ------------------------------------------------------------------
 
-def _clean_breakdown_sentinels(result: dict) -> None:
-    """Replace breakdown sentinel strings in API response results in-place.
-
-    Handles both list-of-lists (materialized/HogQL) and list-of-dicts (inline insight).
-    If the "Other" sentinel is found, it means the breakdown_limit was exceeded —
-    we clean it to "Other" but also capture_exception for visibility.
-    """
-    rows = result.get("results")
-    if not rows:
-        return
-
-    found_other = False
-    columns = result.get("columns")
-    if columns:
-        # HogQL/materialized path: results are list[tuple] or list[list]
-        indices = [i for i, col in enumerate(columns) if col == "breakdown_value"]
-        if not indices:
-            return
-        for row_idx, row in enumerate(rows):
-            needs_clean = any(isinstance(row[i], (str, list)) or row[i] is None for i in indices)
-            if not needs_clean:
-                continue
-            if not found_other:
-                found_other = any(_value_contains_other(row[i], BREAKDOWN_OTHER_STRING_LABEL) for i in indices)
-            row_list = list(row)
-            for i in indices:
-                row_list[i] = _clean_sentinel_value(row_list[i])
-            rows[row_idx] = type(row)(row_list)
-    elif isinstance(rows[0], dict):
-        # Inline insight path: results are list[dict]
-        for row in rows:
-            if "breakdown_value" in row:
-                if not found_other:
-                    found_other = _value_contains_other(row["breakdown_value"], BREAKDOWN_OTHER_STRING_LABEL)
-                row["breakdown_value"] = _clean_sentinel_value(row["breakdown_value"])
-            if "label" in row:
-                if not found_other and isinstance(row["label"], str):
-                    found_other = BREAKDOWN_OTHER_STRING_LABEL in row["label"]
-                row["label"] = _clean_sentinel_label(row["label"])
-
-    if found_other:
-        capture_exception(
-            Exception(
-                f"Endpoint breakdown limit ({ENDPOINT_BREAKDOWN_LIMIT}) exceeded — 'Other' bucket appeared in results"
-            )
+    @extend_schema(exclude=True)
+    @action(methods=["POST"], detail=False)
+    def bulk_update_tags(self, request: Request, *args, **kwargs) -> Response:
+        # The inherited TaggedItemViewSetMixin.bulk_update_tags assumes integer PKs (its
+        # BulkUpdateTagsRequestSerializer validates ids as IntegerField). Endpoint uses
+        # UUID PKs, so the action is unusable. Return 405 until the mixin gains UUID support.
+        return Response(
+            {"detail": "Bulk tag updates are not supported for endpoints."},
+            status=status.HTTP_405_METHOD_NOT_ALLOWED,
         )
-
-
-def _value_contains_other(value: object, other_label: str) -> bool:
-    """Check if a breakdown value contains the 'Other' sentinel."""
-    if isinstance(value, str):
-        return value == other_label
-    if isinstance(value, list):
-        return any(item == other_label for item in value if isinstance(item, str))
-    return False
-
-
-def _clean_sentinel_value(value: object) -> object:
-    """Clean breakdown sentinel strings (null and other) in a value or list of values."""
-    if isinstance(value, str):
-        if value == BREAKDOWN_NULL_STRING_LABEL or value == "":
-            return None
-        if value == BREAKDOWN_OTHER_STRING_LABEL:
-            return "Other"
-    elif isinstance(value, list):
-        return [_clean_sentinel_value(item) for item in value]
-    return value
-
-
-def _clean_sentinel_label(label: object) -> object:
-    """Clean a label string containing ::-joined breakdown parts."""
-    if not isinstance(label, str):
-        return label
-    if BREAKDOWN_NULL_STRING_LABEL not in label and BREAKDOWN_OTHER_STRING_LABEL not in label:
-        return label
-    parts = label.split("::")
-    cleaned = [
-        "null" if p == BREAKDOWN_NULL_STRING_LABEL else "Other" if p == BREAKDOWN_OTHER_STRING_LABEL else p
-        for p in parts
-    ]
-    return "::".join(cleaned)
