+++ products/endpoints/backend/services/validation.py
@@ -1,0 +1,236 @@
+"""Request payload validation for endpoint create/update.
+
+These functions validate (and normalize) the ``EndpointRequest`` payload before
+it reaches the CRUD service: name format, data freshness buckets, HogQL
+parseability, and variable placeholder <-> definition consistency.
+"""
+
+import re
+import uuid
+from typing import Optional
+
+from rest_framework.exceptions import ValidationError
+
+from posthog.schema import EndpointRequest, HogQLQuery, HogQLVariable
+
+from posthog.hogql import ast
+from posthog.hogql.errors import ExposedHogQLError, ResolutionError
+from posthog.hogql.parser import parse_select
+
+from posthog.exceptions_capture import capture_exception
+from posthog.models import Team
+
+from products.endpoints.backend.constants import (
+    DATA_FRESHNESS_BUCKET_SPECS,
+    ENDPOINT_NAME_REGEX,
+    VALID_DATA_FRESHNESS_SECONDS,
+)
+from products.endpoints.backend.materialization_transforms import SUPPORTED_BUCKET_FUNCTIONS, VariablePlaceholderFinder
+from products.endpoints.backend.models import Endpoint, can_materialize_query
+from products.product_analytics.backend.models.insight_variable import InsightVariable
+
+
+def validate_data_freshness(data_freshness_seconds: int | None) -> None:
+    """Validate data_freshness_seconds is one of the allowed bucket values."""
+    if data_freshness_seconds is None:
+        return
+    if data_freshness_seconds not in VALID_DATA_FRESHNESS_SECONDS:
+        allowed = sorted(VALID_DATA_FRESHNESS_SECONDS)
+        human = ", ".join(b.human for b in DATA_FRESHNESS_BUCKET_SPECS)
+        raise ValidationError(
+            {"data_freshness_seconds": f"Data freshness must be one of: {allowed} seconds ({human})."}
+        )
+
+
+def validate_bucket_overrides(bucket_overrides: dict[str, str] | None) -> None:
+    """Raise ValidationError if any bucket override value is not in SUPPORTED_BUCKET_FUNCTIONS."""
+    if not bucket_overrides:
+        return
+
+    invalid = {k: v for k, v in bucket_overrides.items() if v not in SUPPORTED_BUCKET_FUNCTIONS}
+    if invalid:
+        valid_options = list(SUPPORTED_BUCKET_FUNCTIONS.keys())
+        raise ValidationError(f"Invalid bucket override values: {invalid}. Valid options: {valid_options}")
+
+
+def validate_hogql_query(query: HogQLQuery, team: Team) -> None:
+    """Validate that a HogQL query can be parsed and the variables are valid."""
+    try:
+        ast_node = parse_select(query.query)
+    except ExposedHogQLError as e:
+        raise ValidationError({"query": f"Invalid HogQL query: {e}"}) from e
+    except ResolutionError as e:
+        capture_exception(e)
+        raise ValidationError({"query": "Invalid HogQL query: unable to resolve table or field references."})
+    except Exception as e:
+        capture_exception(e)
+        raise ValidationError({"query": "Unknown error occurred parsing the query."})
+
+    validate_variable_placeholders(ast_node, query.variables or {}, team)
+
+
+def sync_hogql_query_variables(query: HogQLQuery, team: Team) -> None:
+    """Sync query variable definitions with the placeholders used in the query string."""
+    try:
+        ast_node = parse_select(query.query)
+    except Exception:
+        # Validation runs separately and will surface parse errors.
+        return
+
+    finder = VariablePlaceholderFinder()
+    finder.visit(ast_node)
+
+    placeholder_names = {str(p.chain[1]) for p in finder.variable_placeholders if p.chain and len(p.chain) > 1}
+    if not placeholder_names:
+        query.variables = None
+        return
+
+    existing_variables = query.variables or {}
+    existing_variable_ids: list[str] = []
+    for variable in existing_variables.values():
+        if not variable.variableId:
+            continue
+        try:
+            uuid.UUID(variable.variableId)
+        except ValueError:
+            continue
+        existing_variable_ids.append(variable.variableId)
+
+    team_variables = InsightVariable.objects.filter(team=team, id__in=existing_variable_ids)
+    team_variables_by_id = {str(variable.id): variable for variable in team_variables}
+
+    synced_variables: dict[str, HogQLVariable] = {}
+    existing_code_names: set[str] = set()
+
+    for variable_id, variable in existing_variables.items():
+        if variable.code_name and variable.code_name in placeholder_names:
+            if variable.value is None and variable.isNull is not True:
+                team_variable = team_variables_by_id.get(variable.variableId)
+                if team_variable:
+                    variable.value = team_variable.default_value
+                    variable.isNull = team_variable.default_value is None
+            synced_variables[variable_id] = variable
+            existing_code_names.add(variable.code_name)
+
+    missing_code_names = placeholder_names - existing_code_names
+    if missing_code_names:
+        missing_variables = InsightVariable.objects.filter(team=team, code_name__in=missing_code_names)
+
+        for variable in missing_variables:
+            if not variable.code_name:
+                continue
+            synced_variables[str(variable.id)] = HogQLVariable(
+                variableId=str(variable.id),
+                code_name=variable.code_name,
+                value=variable.default_value,
+                isNull=variable.default_value is None,
+            )
+
+    query.variables = synced_variables or None
+
+
+def validate_variable_placeholders(node: ast.AST, variables: Optional[dict[str, HogQLVariable]], team: Team) -> None:
+    """Validate that every {variables.X} placeholder in the query has a matching variable definition."""
+    finder = VariablePlaceholderFinder()
+    finder.visit(node)
+
+    if not finder.variable_placeholders:
+        return
+
+    placeholder_names = {str(p.chain[1]) for p in finder.variable_placeholders if p.chain and len(p.chain) > 1}
+
+    defined_code_names: set[str] = set()
+    variable_ids: set[str] = set()
+    if variables:
+        defined_code_names = {v.code_name for v in variables.values() if v.code_name}
+        variable_ids = set(variables.keys())
+
+    undefined: list[str] = sorted(placeholder_names - defined_code_names)
+    if undefined:
+        raise ValidationError(
+            {
+                "query": f"Query references undefined variable(s): {', '.join(undefined)}. "
+                "See https://posthog.com/docs/endpoints/variables for detail."
+            }
+        )
+
+    if variable_ids:
+        valid_uuids: set[str] = set()
+        invalid_uuids: set[str] = set()
+        invalid_ids: set[str] = set()
+        for vid in variable_ids:
+            try:
+                uuid.UUID(vid)
+                valid_uuids.add(vid)
+            except ValueError:
+                invalid_uuids.add(vid)
+
+        if invalid_uuids:
+            raise ValidationError({"query": f"Variable ID(s) not valid UUIDs: {', '.join(sorted(invalid_uuids))}. "})
+
+        if valid_uuids:
+            existing_ids = {
+                str(id)
+                for id in InsightVariable.objects.filter(team=team, id__in=valid_uuids).values_list("id", flat=True)
+            }
+            invalid_ids = valid_uuids - existing_ids
+
+        if invalid_ids:
+            raise ValidationError(
+                {
+                    "query": f"Variable ID(s) not found: {', '.join(sorted(invalid_ids))}. "
+                    "Make sure the variables exist in https://app.posthog.com/data-management/variables."
+                }
+            )
+
+
+def validate_endpoint_request(data: EndpointRequest, team: Team, strict: bool = True) -> None:
+    """Validate a create/update payload. With strict=True, name and query are required."""
+    query = data.query
+    if not query and strict:
+        raise ValidationError({"query": "This field is required."})
+
+    name = data.name
+    if not name:
+        if name is not None or strict:
+            raise ValidationError({"name": "This field is required."})
+        return
+    if not isinstance(name, str) or not re.fullmatch(ENDPOINT_NAME_REGEX, name):
+        raise ValidationError(
+            {
+                "name": f"Invalid name '{name}'. Must start with a letter, contain only alphanumeric characters, "
+                "hyphens, or underscores, and be between 1 and 128 characters long."
+            }
+        )
+
+    if query and isinstance(query, HogQLQuery) and query.query:
+        sync_hogql_query_variables(query, team)
+        validate_hogql_query(query, team)
+
+    validate_data_freshness(data.data_freshness_seconds)
+
+
+def validate_update_request(data: EndpointRequest, team: Team, endpoint: Endpoint | None = None) -> None:
+    """Validate an update payload against the endpoint's resulting state."""
+    validate_data_freshness(data.data_freshness_seconds)
+
+    # Determine final states after this request (for validation)
+    will_be_active = data.is_active if data.is_active is not None else (endpoint.is_active if endpoint else True)
+
+    if not will_be_active and data.is_materialized is True:
+        raise ValidationError({"is_materialized": "Cannot enable materialization on inactive endpoint."})
+
+    if data.is_materialized is True:
+        # Fail fast on queries that can't be materialized. The service re-checks against
+        # the final version (the authoritative guard); this catches it before any writes.
+        effective_query = (
+            data.query.model_dump() if data.query is not None else (endpoint.get_version().query if endpoint else None)
+        )
+        if effective_query is not None:
+            can_materialize, reason = can_materialize_query(effective_query)
+            if not can_materialize:
+                raise ValidationError(f"Cannot materialize endpoint. Reason: {reason}")
+
+    if data.query and isinstance(data.query, HogQLQuery) and data.query.query:
+        sync_hogql_query_variables(data.query, team)
+        validate_hogql_query(data.query, team)
