+++ products/endpoints/backend/tests/test_endpoint_materialization.py
@@ -24,10 +24,11 @@ from posthog.sync import database_sync_to_async
 from products.data_modeling.backend.models.datawarehouse_saved_query import DataWarehouseSavedQuery
 from products.data_modeling.backend.models.modeling import DataWarehouseModelPath
 from products.data_warehouse.backend.data_load.saved_query_service import get_saved_query_schedule
-from products.endpoints.backend.api import EndpointViewSet
-from products.endpoints.backend.materialization import build_endpoint_hogql
+from products.endpoints.backend.materialization_transforms import build_endpoint_hogql
 from products.endpoints.backend.models import EndpointVersion
-from products.endpoints.backend.services.endpoint_materialization_service import (
+from products.endpoints.backend.services.execution import EndpointExecutionService
+from products.endpoints.backend.services.materialization import (
+    EndpointMaterializationService,
     OrphanedEndpointSavedQueryError,
     prepare_executable_query,
 )
@@ -627,7 +628,9 @@ class TestEndpointMaterialization(ClickhouseTestMixin, APIBaseTest):
         )
         saved_query.save()
 
-        with mock.patch("products.endpoints.backend.api.EndpointViewSet._execute_query_and_respond") as mock_execute:
+        with mock.patch(
+            "products.endpoints.backend.services.execution.EndpointExecutionService._execute_query_and_respond"
+        ) as mock_execute:
             old_cache_time = timezone.now() - timedelta(minutes=30)
             old_cached_response = Response(
                 {
@@ -670,7 +673,9 @@ class TestEndpointMaterialization(ClickhouseTestMixin, APIBaseTest):
         )
         new_saved_query.save()
 
-        with mock.patch("products.endpoints.backend.api.EndpointViewSet._execute_query_and_respond") as mock_execute:
+        with mock.patch(
+            "products.endpoints.backend.services.execution.EndpointExecutionService._execute_query_and_respond"
+        ) as mock_execute:
             new_cache_time = timezone.now() - timedelta(minutes=5)
             new_cached_response = Response(
                 {
@@ -777,9 +782,11 @@ class TestEndpointMaterialization(ClickhouseTestMixin, APIBaseTest):
         # Mock the execution methods to track which path is taken
         with (
             mock.patch.object(
-                EndpointViewSet, "_execute_materialized_endpoint", return_value=Response({})
+                EndpointExecutionService, "_execute_materialized_endpoint", return_value=Response({})
             ) as mock_materialized,
-            mock.patch.object(EndpointViewSet, "_execute_inline_endpoint", return_value=Response({})) as mock_inline,
+            mock.patch.object(
+                EndpointExecutionService, "_execute_inline_endpoint", return_value=Response({})
+            ) as mock_inline,
         ):
             response = self.client.post(
                 f"/api/environments/{self.team.id}/endpoints/{endpoint.name}/run",
@@ -828,9 +835,11 @@ class TestEndpointMaterialization(ClickhouseTestMixin, APIBaseTest):
         # Mock the execution methods to track which path is taken
         with (
             mock.patch.object(
-                EndpointViewSet, "_execute_materialized_endpoint", return_value=Response({})
+                EndpointExecutionService, "_execute_materialized_endpoint", return_value=Response({})
             ) as mock_materialized,
-            mock.patch.object(EndpointViewSet, "_execute_inline_endpoint", return_value=Response({})) as mock_inline,
+            mock.patch.object(
+                EndpointExecutionService, "_execute_inline_endpoint", return_value=Response({})
+            ) as mock_inline,
         ):
             response = self.client.post(
                 f"/api/environments/{self.team.id}/endpoints/{endpoint.name}/run",
@@ -877,9 +886,11 @@ class TestEndpointMaterialization(ClickhouseTestMixin, APIBaseTest):
 
         with (
             mock.patch.object(
-                EndpointViewSet, "_execute_materialized_endpoint", return_value=Response({})
+                EndpointExecutionService, "_execute_materialized_endpoint", return_value=Response({})
             ) as mock_materialized,
-            mock.patch.object(EndpointViewSet, "_execute_inline_endpoint", return_value=Response({})) as mock_inline,
+            mock.patch.object(
+                EndpointExecutionService, "_execute_inline_endpoint", return_value=Response({})
+            ) as mock_inline,
         ):
             response = self.client.post(
                 f"/api/environments/{self.team.id}/endpoints/{endpoint.name}/run",
@@ -926,9 +937,11 @@ class TestEndpointMaterialization(ClickhouseTestMixin, APIBaseTest):
 
         with (
             mock.patch.object(
-                EndpointViewSet, "_execute_materialized_endpoint", return_value=Response({})
+                EndpointExecutionService, "_execute_materialized_endpoint", return_value=Response({})
             ) as mock_materialized,
-            mock.patch.object(EndpointViewSet, "_execute_inline_endpoint", return_value=Response({})) as mock_inline,
+            mock.patch.object(
+                EndpointExecutionService, "_execute_inline_endpoint", return_value=Response({})
+            ) as mock_inline,
         ):
             response = self.client.post(
                 f"/api/environments/{self.team.id}/endpoints/{endpoint.name}/run",
@@ -1035,7 +1048,9 @@ class TestEndpointMaterialization(ClickhouseTestMixin, APIBaseTest):
 
         # Execute the endpoint and verify it uses the materialized path correctly
         # Must provide the breakdown variable (required for security - prevents data leakage)
-        with mock.patch.object(EndpointViewSet, "_execute_query_and_respond", return_value=Response({})) as mock_exec:
+        with mock.patch.object(
+            EndpointExecutionService, "_execute_query_and_respond", return_value=Response({})
+        ) as mock_exec:
             response = self.client.post(
                 f"/api/environments/{self.team.id}/endpoints/{endpoint.name}/run",
                 {"variables": {"$browser": "Chrome"}},
@@ -1055,6 +1070,125 @@ class TestEndpointMaterialization(ClickhouseTestMixin, APIBaseTest):
             # Should filter by the breakdown value using has() for array column
             self.assertIn("has(breakdown_value", query_sql)
             self.assertIn("chrome", query_sql)
+
+    def test_materialized_insight_filters_override_without_value_does_not_bypass_required_breakdown(self):
+        # SECURITY: a filters_override whose property carries no usable value adds no WHERE clause
+        # in apply_materialized_filters, so it must not be allowed to skip the required-breakdown
+        # check — otherwise the caller gets every materialized breakdown row.
+        trends_query_with_breakdown = {
+            "kind": "TrendsQuery",
+            "series": [{"kind": "EventsNode", "event": "$pageview", "math": "total"}],
+            "dateRange": {"date_from": "-7d"},
+            "interval": "day",
+            "breakdownFilter": {"breakdown": "$browser", "breakdown_type": "event", "breakdown_limit": 5},
+        }
+        _create_event(team=self.team, event="$pageview", distinct_id="user1")
+        flush_persons_and_events()
+
+        endpoint = create_endpoint_with_version(
+            name="trends_breakdown_secure",
+            team=self.team,
+            query=trends_query_with_breakdown,
+            created_by=self.user,
+            is_active=True,
+        )
+        response = self.client.patch(
+            f"/api/environments/{self.team.id}/endpoints/{endpoint.name}/",
+            {"is_materialized": True, "data_freshness_seconds": 43200},
+            format="json",
+        )
+        self.assertEqual(response.status_code, status.HTTP_200_OK, response.json())
+
+        version = endpoint.versions.first()
+        version.refresh_from_db()
+        saved_query = version.saved_query
+        assert saved_query is not None
+        saved_query.status = DataWarehouseSavedQuery.Status.COMPLETED
+        saved_query.last_run_at = timezone.now()
+        saved_query.table = DataWarehouseTable.objects.create(
+            team=self.team,
+            name="trends_breakdown_secure",
+            format=DataWarehouseTable.TableFormat.Parquet,
+            url_pattern="s3://test-bucket/trends-breakdown-secure",
+        )
+        saved_query.save()
+
+        # Property with a key but no value: passes the deprecated-override gate yet yields no filter.
+        with mock.patch.object(EndpointExecutionService, "_execute_query_and_respond", return_value=Response({})):
+            response = self.client.post(
+                f"/api/environments/{self.team.id}/endpoints/{endpoint.name}/run",
+                {"filters_override": {"properties": [{"key": "$browser", "type": "event"}]}},
+                format="json",
+            )
+        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST, response.json())
+        self.assertIn("Required variable", str(response.json()))
+
+        # A filters_override that DOES carry a usable value still satisfies the requirement.
+        with mock.patch.object(EndpointExecutionService, "_execute_query_and_respond", return_value=Response({})):
+            response = self.client.post(
+                f"/api/environments/{self.team.id}/endpoints/{endpoint.name}/run",
+                {"filters_override": {"properties": [{"key": "$browser", "type": "event", "value": "Chrome"}]}},
+                format="json",
+            )
+        self.assertEqual(response.status_code, status.HTTP_200_OK, response.json())
+
+    def test_materialized_multi_breakdown_filters_override_does_not_bypass_required_breakdowns(self):
+        # SECURITY: filters_override applies only its first property as a single breakdown filter, so
+        # it can never satisfy a multi-breakdown endpoint — supplying one value must not let the caller
+        # omit the other required breakdown variables (which would leak rows across them).
+        trends_multi_breakdown = {
+            "kind": "TrendsQuery",
+            "series": [{"kind": "EventsNode", "event": "$pageview", "math": "total"}],
+            "dateRange": {"date_from": "-7d"},
+            "interval": "day",
+            "breakdownFilter": {
+                "breakdowns": [
+                    {"property": "$browser", "type": "event"},
+                    {"property": "$os", "type": "event"},
+                ],
+                "breakdown_limit": 5,
+            },
+        }
+        _create_event(team=self.team, event="$pageview", distinct_id="user1")
+        flush_persons_and_events()
+
+        endpoint = create_endpoint_with_version(
+            name="trends_multi_breakdown_secure",
+            team=self.team,
+            query=trends_multi_breakdown,
+            created_by=self.user,
+            is_active=True,
+        )
+        response = self.client.patch(
+            f"/api/environments/{self.team.id}/endpoints/{endpoint.name}/",
+            {"is_materialized": True, "data_freshness_seconds": 43200},
+            format="json",
+        )
+        self.assertEqual(response.status_code, status.HTTP_200_OK, response.json())
+
+        version = endpoint.versions.first()
+        version.refresh_from_db()
+        saved_query = version.saved_query
+        assert saved_query is not None
+        saved_query.status = DataWarehouseSavedQuery.Status.COMPLETED
+        saved_query.last_run_at = timezone.now()
+        saved_query.table = DataWarehouseTable.objects.create(
+            team=self.team,
+            name="trends_multi_breakdown_secure",
+            format=DataWarehouseTable.TableFormat.Parquet,
+            url_pattern="s3://test-bucket/trends-multi-breakdown-secure",
+        )
+        saved_query.save()
+
+        # Supplying only one breakdown value via filters_override leaves $os unfiltered.
+        with mock.patch.object(EndpointExecutionService, "_execute_query_and_respond", return_value=Response({})):
+            response = self.client.post(
+                f"/api/environments/{self.team.id}/endpoints/{endpoint.name}/run",
+                {"filters_override": {"properties": [{"key": "$browser", "type": "event", "value": "Chrome"}]}},
+                format="json",
+            )
+        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST, response.json())
+        self.assertIn("Required variable", str(response.json()))
 
     def test_materialized_hogql_endpoint_with_variable_executes_correctly(self):
         """Test that HogQL endpoints with variables work when materialized.
@@ -1130,7 +1264,9 @@ class TestEndpointMaterialization(ClickhouseTestMixin, APIBaseTest):
         saved_query.save()
 
         # Execute with variable filter
-        with mock.patch.object(EndpointViewSet, "_execute_query_and_respond", return_value=Response({})) as mock_exec:
+        with mock.patch.object(
+            EndpointExecutionService, "_execute_query_and_respond", return_value=Response({})
+        ) as mock_exec:
             response = self.client.post(
                 f"/api/environments/{self.team.id}/endpoints/{endpoint.name}/run",
                 {"variables": {"event_name": "$pageview"}},
@@ -1352,7 +1488,9 @@ class TestEndpointMaterialization(ClickhouseTestMixin, APIBaseTest):
         # Update with a new query — version creation succeeds, but materialization fails
         new_query = {"kind": "HogQLQuery", "query": "SELECT * FROM events WHERE timestamp > now() - INTERVAL 1 DAY"}
         with mock.patch.object(
-            EndpointViewSet, "_enable_materialization_inner", side_effect=Exception("Temporal unavailable")
+            EndpointMaterializationService,
+            "_enable_materialization_inner",
+            side_effect=Exception("Temporal unavailable"),
         ):
             response = self.client.put(
                 f"/api/environments/{self.team.id}/endpoints/{endpoint.name}/",
@@ -1380,7 +1518,9 @@ class TestEndpointMaterialization(ClickhouseTestMixin, APIBaseTest):
         )
 
         with mock.patch.object(
-            EndpointViewSet, "_enable_materialization_inner", side_effect=Exception("Temporal unavailable")
+            EndpointMaterializationService,
+            "_enable_materialization_inner",
+            side_effect=Exception("Temporal unavailable"),
         ):
             response = self.client.patch(
                 f"/api/environments/{self.team.id}/endpoints/{endpoint.name}/",
@@ -1388,7 +1528,8 @@ class TestEndpointMaterialization(ClickhouseTestMixin, APIBaseTest):
                 format="json",
             )
 
-        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
+        # Unexpected internal failures surface as a server error, not a request error
+        self.assertEqual(response.status_code, status.HTTP_500_INTERNAL_SERVER_ERROR)
 
     def test_materialized_relative_date_range_advances_on_refresh(self):
         _create_event(team=self.team, event="$pageview", distinct_id="u1")
@@ -1494,6 +1635,83 @@ class TestEndpointMaterialization(ClickhouseTestMixin, APIBaseTest):
         write_prefixes = ("insert", "update", "delete")
         writes = [q["sql"] for q in ctx.captured_queries if q["sql"].lstrip().lower().startswith(write_prefixes)]
         self.assertEqual(writes, [], f"build_endpoint_hogql wrote to Postgres: {writes}")
+
+    def test_deactivating_one_version_keeps_current_version_materialization(self):
+        """PATCH ?version=N {is_active: false} must not touch the current version's materialization."""
+        endpoint = create_endpoint_with_version(
+            name="version-deactivate-endpoint",
+            team=self.team,
+            query=self.sample_hogql_query,
+            created_by=self.user,
+            is_active=True,
+        )
+
+        # Create v2 by changing the query
+        response = self.client.patch(
+            f"/api/environments/{self.team.id}/endpoints/{endpoint.name}/",
+            {"query": {"kind": "HogQLQuery", "query": "SELECT event FROM events LIMIT 5"}},
+            format="json",
+        )
+        self.assertEqual(response.status_code, status.HTTP_200_OK, response.json())
+
+        # Materialize v2 (the current version)
+        response = self.client.patch(
+            f"/api/environments/{self.team.id}/endpoints/{endpoint.name}/",
+            {"is_materialized": True},
+            format="json",
+        )
+        self.assertEqual(response.status_code, status.HTTP_200_OK, response.json())
+        v2 = endpoint.get_version(2)
+        self.assertIsNotNone(v2.saved_query)
+
+        # Deactivate v1 only
+        response = self.client.patch(
+            f"/api/environments/{self.team.id}/endpoints/{endpoint.name}/",
+            {"is_active": False, "version": 1},
+            format="json",
+        )
+        self.assertEqual(response.status_code, status.HTTP_200_OK, response.json())
+
+        v1 = endpoint.get_version(1)
+        self.assertFalse(v1.is_active)
+        endpoint.refresh_from_db()
+        self.assertTrue(endpoint.is_active)
+
+        # v2's materialization must be untouched
+        v2.refresh_from_db()
+        self.assertIsNotNone(v2.saved_query_id, "current version's materialization was disabled by mistake")
+        assert v2.saved_query is not None
+        self.assertFalse(v2.saved_query.deleted)
+
+    def test_enable_materialization_does_not_hijack_user_saved_query(self):
+        """A user-created saved query whose name collides with {endpoint}_v{n} must not be taken over."""
+        endpoint = create_endpoint_with_version(
+            name="hijack-target",
+            team=self.team,
+            query=self.sample_hogql_query,
+            created_by=self.user,
+            is_active=True,
+        )
+        user_query = {"kind": "HogQLQuery", "query": "SELECT 42 AS answer"}
+        user_saved_query = DataWarehouseSavedQuery.objects.create(
+            team=self.team,
+            name="hijack-target_v1",
+            query=user_query,
+        )
+
+        response = self.client.patch(
+            f"/api/environments/{self.team.id}/endpoints/{endpoint.name}/",
+            {"is_materialized": True},
+            format="json",
+        )
+
+        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST, response.json())
+
+        user_saved_query.refresh_from_db()
+        self.assertEqual(user_saved_query.query, user_query, "user's saved query was overwritten")
+        self.assertFalse(user_saved_query.is_materialized)
+        version = endpoint.get_version()
+        self.assertIsNone(version.saved_query_id)
 
 
 @pytest.mark.asyncio
