+++ products/endpoints/backend/tests/test_variable_materialization.py
@@ -8,7 +8,7 @@ from parameterized import parameterized
 from posthog.hogql import ast
 from posthog.hogql.parser import parse_select
 
-from products.endpoints.backend.materialization import (
+from products.endpoints.backend.materialization_transforms import (
     DownstreamCTEShape,
     _build_cte_read_graph,
     _classify_downstream_cte,
@@ -747,7 +747,7 @@ class TestCombinatorReaggregation(APIBaseTest):
         ]
     )
     def test_reaggregatable_combinators_allowed(self, func_name, expected_reagg):
-        from products.endpoints.backend.materialization import get_reaggregation
+        from products.endpoints.backend.materialization_transforms import get_reaggregation
 
         reagg = get_reaggregation(func_name)
         assert reagg is not None, f"{func_name} should be re-aggregatable"
@@ -767,7 +767,7 @@ class TestCombinatorReaggregation(APIBaseTest):
         ]
     )
     def test_non_reaggregatable_functions_rejected(self, func_name):
-        from products.endpoints.backend.materialization import get_reaggregation
+        from products.endpoints.backend.materialization_transforms import get_reaggregation
 
         reagg = get_reaggregation(func_name)
         assert reagg is None, f"{func_name} should NOT be re-aggregatable"
@@ -819,7 +819,7 @@ class TestStripCombinators(APIBaseTest):
         ]
     )
     def test_strips_to_known_base(self, func_name, expected_base):
-        from products.endpoints.backend.materialization import _strip_combinators
+        from products.endpoints.backend.materialization_transforms import _strip_combinators
 
         assert _strip_combinators(func_name) == expected_base
 
@@ -835,13 +835,13 @@ class TestStripCombinators(APIBaseTest):
         ]
     )
     def test_returns_none_for_unknown(self, func_name):
-        from products.endpoints.backend.materialization import _strip_combinators
+        from products.endpoints.backend.materialization_transforms import _strip_combinators
 
         result = _strip_combinators(func_name)
         # Should return the base but it won't be in REAGGREGATABLE_BASE_FUNCTIONS
         # For truly unknown functions, returns None
         if result is not None:
-            from products.endpoints.backend.materialization import REAGGREGATABLE_BASE_FUNCTIONS
+            from products.endpoints.backend.materialization_transforms import REAGGREGATABLE_BASE_FUNCTIONS
 
             # The base was found but it's not in the registry — that's the expected path
             # for functions like uniq, avg whose base is known but not re-aggregatable
@@ -1407,7 +1407,7 @@ class TestMaterializedQueryExecution(APIBaseTest):
         assert True  # See _transform_select_for_materialized_table implementation
 
     def test_select_transformation_with_alias(self):
-        from products.endpoints.backend.materialization import transform_select_for_materialized_table
+        from products.endpoints.backend.materialization_transforms import transform_select_for_materialized_table
 
         query_str = "SELECT count() as total, toStartOfDay(timestamp) as date FROM events"
         parsed = parse_select(query_str)
@@ -1429,7 +1429,7 @@ class TestMaterializedQueryExecution(APIBaseTest):
         assert transformed[1].is_aggregate is False
 
     def test_select_transformation_without_alias(self):
-        from products.endpoints.backend.materialization import transform_select_for_materialized_table
+        from products.endpoints.backend.materialization_transforms import transform_select_for_materialized_table
 
         query_str = "SELECT count() FROM events"
         parsed = parse_select(query_str)
@@ -1684,7 +1684,7 @@ class TestTransformQuerySnapshots(APIBaseTest):
         assert group_by_columns.count("event") == 1, f"GROUP BY has duplicate 'event': {group_by_columns}"
 
     def test_ast_node_not_shared_between_select_and_group_by(self):
-        from products.endpoints.backend.materialization import MaterializationTransformer
+        from products.endpoints.backend.materialization_transforms import MaterializationTransformer
 
         query_str = "SELECT count() FROM events WHERE toDate(timestamp) >= {variables.from_date}"
         variables = {"var-1": {"code_name": "from_date", "value": "2024-01-01"}}
@@ -1721,7 +1721,7 @@ class TestMaterializedReadPath(APIBaseTest):
 
     def _build_read_query(self, query_str: str, variables_meta: dict, variable_values: dict) -> str:
         """Simulate the materialized read path: analyze variables, then build a SELECT with filters."""
-        from products.endpoints.backend.api import EndpointViewSet
+        from products.endpoints.backend.services.strategies import apply_where_filter
 
         hogql_query = {"kind": "HogQLQuery", "query": query_str, "variables": variables_meta}
         _, _, var_infos = analyze_variables_for_materialization(hogql_query)
@@ -1731,11 +1731,10 @@ class TestMaterializedReadPath(APIBaseTest):
             select_from=ast.JoinExpr(table=ast.Field(chain=["materialized_table"])),
         )
 
-        viewset = EndpointViewSet()
         for mat_var in var_infos:
             var_value = variable_values.get(mat_var.code_name)
             if var_value is not None:
-                viewset._apply_where_filter(
+                apply_where_filter(
                     select_query,
                     mat_var.code_name,
                     var_value,
@@ -2609,7 +2608,7 @@ class TestDownstreamCTEClassifier(APIBaseTest):
     def test_extract_aggregate_name_canonicalizes_count_distinct(self, src, expected):
         from posthog.hogql.parser import parse_expr as _parse_expr
 
-        from products.endpoints.backend.materialization import _extract_aggregate_name as _extract
+        from products.endpoints.backend.materialization_transforms import _extract_aggregate_name as _extract
 
         assert _extract(_parse_expr(src)) == expected
 
@@ -2625,7 +2624,7 @@ class TestDownstreamCTEClassifier(APIBaseTest):
     def test_extract_aggregate_name_canonicalizes_base_aggregates(self, src, expected):
         from posthog.hogql.parser import parse_expr as _parse_expr
 
-        from products.endpoints.backend.materialization import _extract_aggregate_name as _extract
+        from products.endpoints.backend.materialization_transforms import _extract_aggregate_name as _extract
 
         assert _extract(_parse_expr(src)) == expected
