+++ products/signals/backend/models.py
@@ -29,6 +29,7 @@ class SignalSourceConfig(UUIDModel):
         SIGNALS_SCOUT = "signals_scout", "Signals scout"
         LOGS = "logs", "Logs"
         HEALTH_CHECKS = "health_checks", "Health checks"
+        ENDPOINTS = "endpoints", "Endpoints"
 
     class SourceType(models.TextChoices):
         SESSION_ANALYSIS_CLUSTER = "session_analysis_cluster", "Session analysis cluster"
@@ -41,6 +42,8 @@ class SignalSourceConfig(UUIDModel):
         CROSS_SOURCE_ISSUE = "cross_source_issue", "Cross source issue"
         ALERT_STATE_CHANGE = "alert_state_change", "Alert state change"
         HEALTH_ISSUE = "health_issue", "Health issue"
+        ENDPOINT_EXECUTION_FAILED = "endpoint_execution_failed", "Endpoint execution failed"
+        ENDPOINT_BREAKDOWN_LIMIT_EXCEEDED = "endpoint_breakdown_limit_exceeded", "Endpoint breakdown limit exceeded"
 
     team = models.ForeignKey("posthog.Team", on_delete=models.CASCADE, related_name="signal_source_configs")
     source_product = models.CharField(max_length=100, choices=SourceProduct)
